% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_metadata.R
\name{attach_meta}
\alias{attach_meta}
\alias{add_meta}
\title{Attach experimental metadata}
\usage{
attach_meta(peak_table, metadata, column)
}
\arguments{
\item{peak_table}{Peak table}

\item{meta}{Dataframe}
}
\value{
Peak table with metadata added
}
\description{
Attach experimental metadata to peak table.
}
\references{
Eilers, P.H.C. 2004.a
}
\seealso{
\code{\link[ptw:ptw]{ptw}}, \code{\link{correctPeaks}}
}
\author{
Ethan Bass
}
