% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_peak.R
\name{check_peak}
\alias{check_peak}
\alias{compare_spectra}
\title{Function to correct false zeros in peak table.}
\usage{
check_peak(
  peak,
  peak_table,
  chrom_list,
  thresh_auto = 0.95,
  thresh_man = NULL,
  r = 100,
  plot_it = FALSE,
  lambda = "256",
  zeros = FALSE,
  ref = "max",
  order_by = "distance",
  verbose = T,
  plot_diff = TRUE,
  ...
)
}
\arguments{
\item{peak}{Name of peak to be investigated.}

\item{peak_table}{A peak_table object from \code{\link{get_peaktable}}.}

\item{chrom_list}{A list of chromatograms in matrix form (timepoints x
wavelengths).}

\item{thresh_auto}{Defines minimum spectral similarity threshold for
automatic match.}

\item{thresh_man}{Defines minimum spectral similarity threshold to trigger
manual matching process. Should be set equal or lower to `thresh_auto`.}

\item{r}{Defines radius around focal peak to search for matches.}

\item{plot_it}{Logical. If TRUE, plots spectra for comparison.}

\item{lambda}{Wavelength for plot.}

\item{zeros}{Logical. If TRUE, only check zero rows in peak table.}

\item{ref}{Defines chromatogram for reference spectrum. If "max" choose
chromatogram with maximum absorbance.}

\item{order_by}{If "distance", will order matches by distance from
theoretical retention time.}

\item{verbose}{Logical. If TRUE, prints verbose output to console.}

\item{plot_diff}{Logical.}

\item{\dots}{Additional arguments}
}
\value{
A peak table similar to the input peak table, but with changes to
entries selected column as defined by the algorithm.
}
\description{
Function that tries to correct false zeroes for a particular peak in the
peak_table. This can be used as a last resort to automatically (or semi-
automatically) align peak data. In each chromatogram, the function compares
all peaks within a certain radius around the focal peak on the basis of their
spectral similarity to a reference spectrum.
}
\seealso{
\code{\link{get_peaks}}
}
\author{
Ethan Bass
}
