% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatographR-package.R
\docType{package}
\name{chromatographR-package}
\alias{chromatographR}
\alias{chromatographR-package}
\title{chromatographR: Chromatographic Data Analysis Toolset}
\description{
Tools for high-throughput analysis of HPLC-DAD/UV chromatographic data.
The package provides functionality for preprocessing, alignment, peak
detection and fitting, peak table construction, and visualization of
chromatographic data.
}
\details{
A typical workflow includes signal preprocessing and peak table construction
following general approaches described in Wehrens et al. (2015).

Chromatogram alignment is implemented using parametric time warping (PTW) or
variable penalty dynamic time warping (VPdtw).

Peak detection is based on zero-crossings in smoothed derivatives. Peaks are
then modeled using Gaussian, exponential-Gaussian hybrid, or bidirectional
exponentially modified Gaussian peak functions via nonlinear least squares.

Further details and example workflows are provided in the package vignettes.
}
\section{Core analysis functions}{

\itemize{
\item \code{read_chroms()}: Import chromatograms from a variety of vendor formats.
\item \code{preprocess()}: Preprocess chromatographic matrices.
\item \code{correct_rt()}: Align chromatograms.
\item \code{get_peaks()}: Find and integrate peaks.
\item \code{get_peaktable()}: Assemble peak table.
\item \code{attach_metadata()}: Attach metadata to peak table.
\item \code{attach_ref_spectra()}: Attach reference spectra to peak table.
\item \code{normalize_data()}: Normalize \code{peak_table} or \code{chrom_list}.
}
}

\section{Peak refinement}{

\itemize{
\item \code{combine_peaks()}: Combine duplicate peaks based on retention time and spectral similarity.
\item \code{merge_peaks()}: Merge split peaks into a single feature in a peak table.
\item \code{filter_peaktable()}: Filter peak features in a peak table.
\item \code{filter_peaks()}: Filter peaks in peak lists.
}
}

\section{Visualization functions}{

\subsection{Chromatogram plots}{
\itemize{
\item \code{plot_chroms()}: Plot chromatograms as traces.
\item \code{plot_chroms_heatmap()}: Plot chromatograms as heatmap.
\item \code{annotate_peaks()}: Add peak labels to chromatogram plots.
\item \code{mirror_plot()}: Plot chromatograms as mirror plots.
\item \code{plot.ptw_list()}: Plot parametric time warping (PTW) alignment object.
}
}

\subsection{Spectral plots}{
\itemize{
\item \code{plot_spectrum()}: Plot spectrum and/or trace of specified peak.
\item \code{plot_spectrum_inset()}: Plot spectrum over chromatogram.
\item \code{plot_all_spectra()}: Plot all spectra for specified peak.
\item \code{scan_chrom()}: Interactively extract spectra from a chromatogram.
}
}

\subsection{Peak plots}{
\itemize{
\item \code{plot.peak_list()}: Plot fitted peaks over chromatographic trace.
\item \code{plot.peak_table()}: Plot chromatograms and/or spectra from a peak table.
\item \code{boxplot.peak_table()}: Create boxplot from a peak table object.
}
}
}

\section{Data utilities and I/O}{

\itemize{
\item \code{get_times()}: Return retention times from a peak table or a list of chromatograms.
\item \code{get_lambdas()}: Return wavelengths from a peak table or a list of chromatograms.
\item \code{write_peaktable()}: Export peak table to CSV or XLSX format.
\item \code{reshape_chroms()}: Reshape a list of chromatograms to long format.
\item \code{reshape_peaktable()}: Reshape a peak table to long format.
}
}

\section{Example data}{

\itemize{
\item \code{Sa}: A list of four goldenrod root chromatograms.
\item \code{Sa_pr}: Preprocessed goldenrod root chromatograms.
\item \code{Sa_warp}: Preprocessed and aligned goldenrod root chromatograms.
\item \code{pk_tab}: Peak table from aligned goldenrod root chromatograms.
}
}

\references{
\itemize{
\item Clifford, D., Stone, G., Montoliu, I., Rezzi, S., Martin, F. P., Guy, P.,
Bruce, S., & Kochhar, S. 2009. Alignment using variable penalty dynamic time
warping. \emph{Analytical chemistry}, 81(3):1000-1007. \doi{10.1021/ac802041e}.
\item Clifford, D., & Stone, G. 2012. Variable Penalty Dynamic Time Warping Code
for Aligning Mass Spectrometry Chromatograms in R. \emph{Journal of
Statistical Software}, 47(8):1-17. \doi{10.18637/jss.v047.i08}.
\item Eilers, P.H.C. 2004. Parametric Time Warping.
\emph{Analytical Chemistry}, 76:404-411. \doi{10.1021/ac034800e}.
\item Wehrens, R., Bloemberg, T.G., and Eilers P.H.C. 2015. Fast
parametric time warping of peak lists. \emph{Bioinformatics}, 31:3063-3065.
\doi{10.1093/bioinformatics/btv299}.
\item Wehrens, R., Carvalho, E., Fraser, P.D. 2015. Metabolite profiling in
LC–DAD using multivariate curve resolution: the alsace package for R.
\emph{Metabolomics}, 11:143-154. \doi{10.1007/s11306-014-0683-5}.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ethanbass.github.io/chromatographR/}
  \item \url{https://github.com/ethanbass/chromatographR/}
  \item Report bugs at \url{https://github.com/ethanbass/chromatographR/issues/}
}

}
\author{
Ethan Bass
}
