% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_table-methods.R
\name{dim.peak_table}
\alias{dim.peak_table}
\title{Return dimensions of a \code{peak_table} object.}
\usage{
\method{dim}{peak_table}(x)
}
\arguments{
\item{x}{A \code{\link{peak_table}} object.}
}
\value{
Returns the number of rows and columns in \code{peak_table$tab}.
}
\description{
Returns the dimensions of a \code{peak_table}, where the first dimension is
the number of samples and the second dimension is the number of peaks.
}
\keyword{internal}
