% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_gaps.R
\name{fill_gaps}
\alias{fill_gaps}
\title{Fill gaps in peak table}
\usage{
fill_gaps(
  peak_table,
  chrom_list,
  ref = c("max.int", "max.cor"),
  similarity_threshold = 0.95,
  rt_tolerance = 0.2,
  spectral_weight = 0.5,
  only_zeros = FALSE,
  peaks_only = TRUE,
  plot_it = FALSE
)
}
\arguments{
\item{peak_table}{A \code{peak_table} object.}

\item{chrom_list}{A list of chromatograms in matrix format}

\item{ref}{How to select reference spectrum: either maximum correlation
(\code{max.cor}) or maximum intensity (\code{max.int}).}

\item{similarity_threshold}{Minimum spectral similarity.}

\item{rt_tolerance}{Maximum retention time difference.}

\item{spectral_weight}{Relative weight of spectral similarity and retention
time similarity.}

\item{only_zeros}{Only consider zeros in peak_table for filling.}

\item{peaks_only}{Only consider peaks found by \code{\link{find_peaks}}.}

\item{plot_it}{Whether to plot difference between filled and unfilled peak tables.}
}
\description{
Fill gaps in peak table
}
\author{
Ethan Bass
}
