% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_peak.R
\name{gather_reference_spectra}
\alias{gather_reference_spectra}
\title{Function to gather reference spectra.}
\usage{
gather_reference_spectra(
  peak_table,
  chrom_list,
  ref_criteria = c("max.cor", "max.sig")
)
}
\arguments{
\item{peak_table}{Peak table from \code{\link{getPeakTable}}.}

\item{chrom_list}{A list of chromatograms in matrix form (timepoints x
wavelengths).}

\item{ref_criteria}{What criterion to use to select reference spectra.
Current options are maximum correlation ("max.cor") or maximum signal
intensity ("max.sig").}

\item{\dots}{Additional arguments}
}
\value{
A matrix consisting of reference spectra for each peak in the
provided peak table.
}
\description{
Function that tries to correct false zeroes for a particular peak in the
peaktable. In each chromatogram, the function compares all peaks within a
certain radius around the focal peak on the basis of their spectral
similarity to a reference spectrum.
}
\seealso{
\code{\link{get_peaks}}
}
\author{
Ethan Bass
}
