% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_lambdas}
\alias{get_lambdas}
\title{Get lambdas}
\usage{
get_lambdas(x)
}
\arguments{
\item{x}{A list of chromatograms or \code{peak_table} object.}
}
\value{
A numeric vector of wavelengths.
}
\description{
Get wavelengths from a list of chromatograms or a \code{peak_table} object.
}
\examples{
data(Sa_warp)
get_lambdas(Sa_warp)
}
\seealso{
Other utility functions:
\code{\link[=combine_peaks]{combine_peaks()}},
\code{\link[=get_times]{get_times()}},
\code{\link[=merge_peaks]{merge_peaks()}},
\code{\link[=reshape_chroms]{reshape_chroms()}},
\code{\link[=reshape_peaktable]{reshape_peaktable()}}
}
\concept{utility functions}
\keyword{internal}
