% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_purity.R, R/peak_purity.R
\name{get_mean_purity}
\alias{get_mean_purity}
\title{Calculate mean peak purity}
\usage{
get_mean_purity(x, pos, weight = 1, cutoff = 0.05)

get_mean_purity(x, pos, weight = 1, cutoff = 0.05)
}
\arguments{
\item{x}{A chromatogram in matrix format}

\item{pos}{A vector containing peak information}

\item{weight}{weight provided to \code{\link{get_agilent_threshold}}}

\item{cutoff}{Proportion of maximum absorbance to use as cutoff.
Argument to \code{\link{trim_peak}}.}
}
\description{
Calculate mean peak purity

Calculate mean peak purity
}
\author{
Ethan Bass
}
