% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_purity.R, R/peak_purity.R
\name{get_noise_variance}
\alias{get_noise_variance}
\title{Calculate variance of noise regions}
\usage{
get_noise_variance(x, noise_threshold = 0.005, lambdas = c(210:400))

get_noise_variance(x, noise_threshold = 0.005, lambdas = c(210:400))
}
\arguments{
\item{x}{A chromatogram in matrix format}

\item{noise_threshold}{Threshold to define noise. Highest proportion of maximum absorbance.}

\item{lambdas}{Wavelengths to include}
}
\description{
Calculate variance of noise regions

Calculate variance of noise regions
}
\author{
Ethan Bass
}
