% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_purity.R, R/peak_purity.R
\name{get_purity_values}
\alias{get_purity_values}
\title{Calculate peak purity values}
\usage{
get_purity_values(x, pos, weight = 1)

get_purity_values(x, pos, weight = 1)
}
\arguments{
\item{x}{A chromatogram in matrix format}

\item{pos}{A vector containing peak information}

\item{weight}{weight provided to \code{\link{get_agilent_threshold}}}
}
\description{
Calculate peak purity values

Calculate peak purity values
}
\references{
Stahl, Mark. “Peak Purity Analysis in HPLC and CE Using Diode-Array Technology.”
Agilent Technologies, April 1, 2003, 16.
/href{https://www.agilent.com/cs/library/applications/5988-8647EN.pdf}

Stahl, Mark. “Peak Purity Analysis in HPLC and CE Using Diode-Array Technology.”
Agilent Technologies, April 1, 2003, 16.
/href{https://agilent.com/cs/library/applications/5988-8647EN.pdf}
}
\author{
Ethan Bass
}
