% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_purity.R
\name{get_spectral_similarity}
\alias{get_spectral_similarity}
\title{Calculate spectral similarity}
\usage{
get_spectral_similarity(x, pos)
}
\arguments{
\item{x}{A chromatogram in matrix format}

\item{pos}{A vector containing peak information}
}
\value{
Returns a vector of spectral similarities of the reference spectrum
to the spectrum at each timepoint within the peak specified by \code{pos}.
}
\description{
Calculate spectral similarity
}
\references{
Stahl, Mark. “Peak Purity Analysis in HPLC and CE Using Diode-Array Technology.”
Agilent Technologies, April 1, 2003, 16.
/href{https://www.agilent.com/cs/library/applications/5988-8647EN.pdf}
}
\author{
Ethan Bass
}
