% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_table-methods.R
\name{head.peak_table}
\alias{head.peak_table}
\alias{tail.peak_table}
\title{Return first or last parts of a \code{peak_table}.}
\usage{
\method{head}{peak_table}(x, ...)

\method{tail}{peak_table}(x, ...)
}
\arguments{
\item{x}{A \code{\link{peak_table}} object.}

\item{...}{Additional arguments to \code{\link{tail}}.}
}
\value{
The first or last \code{n} rows of \code{peak_table$tab}.
}
\description{
Returns the first or last parts of the \code{peak_table} object.
}
\keyword{internal}
