% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_chroms.R
\name{load_chroms}
\alias{load_chroms}
\title{Import chromatograms.}
\usage{
load_chroms(
  paths,
  format.in = c("csv", "chemstation", "masshunter"),
  sep = ",",
  dat = NULL,
  ...
)
}
\arguments{
\item{paths}{Path(s) to folders where chromatograms are stored.}

\item{format.in}{Format of files.}

\item{sep}{Argument provided to \code{read.csv}. Defaults to ",".}

\item{dat}{Optional list of chromatograms. If list is provided, the function
will append newly imported chromatograms to the existing list.}

\item{...}{Additional arguments to \code{\link{read.csv}}.}
}
\value{
A list of chromatograms in matrix format.
}
\description{
Convenience function to import chromatograms from a list of folders.
}
\note{
Relies on the file parsers from the
\url{https://github.com/bovee/aston}{Aston} package to import chemstation
\code{.uv} and masshunter \code{.sp} files.
}
\examples{
\dontrun{
###  import from single folder
dat <- load_chromes(paths = path)
### import from multiple folders
path = 'foo'
folders <- list.files(path = path, pattern = "EXPORT3D")
dat <- load_chroms(folders)
}
}
\author{
Ethan Bass
}
