% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{plot_spectrum_inset}
\alias{plot_spectrum_inset}
\title{Plot a chromatogram or chromatograms with an inset UV spectrum}
\usage{
plot_spectrum_inset(
  loc,
  peak_table,
  chrom_list = NULL,
  idx = "max",
  lambda = "max",
  position = "topright",
  inset_width = 0.35,
  inset_height = 0.35,
  ylim = NULL,
  xlim = NULL,
  engine = "ggplot",
  scale_spectrum = TRUE,
  verbose = getOption("verbose"),
  inset_text_size = 8,
  ...
)
}
\arguments{
\item{loc}{Peak name or vector of names in the peak table (e.g. \code{"V10"}).}

\item{peak_table}{A \code{peak_table} object created by \code{\link{get_peaktable}}.}

\item{chrom_list}{A list of chromatograms in matrix format (timepoints x
wavelengths). If no argument is provided here, the function will try to find
the \code{chrom_list} object using the pointer in the \code{peak_table}.}

\item{idx}{Index of chromatogram(s) to plot. If \code{"max"} (default) or a
scalar, plots a single trace using \code{\link[=plot_spectrum]{plot_spectrum()}}. If a vector,
plots multiple traces using \code{\link[=plot_chroms]{plot_chroms()}}.}

\item{lambda}{Wavelength to use for the chromatographic trace and for
locating the peak apex. Either a numeric value or \code{"max"} (default)
to use the wavelength of maximum absorbance.}

\item{position}{Position of the inset. One of \code{"topright"} (default),
\code{"topleft"}, \code{"top"}, \code{"left"}, or \code{"right"}. Alternatively, a numeric
vector of length 2 giving the proportional coordinates of the bottom-left
corner, e.g., \code{c(0.6, 0.6)}, used together with \code{inset_width} and
\code{inset_height}.}

\item{inset_width}{Width of the inset as a proportion of the plot width.
Defaults to \code{0.35}.}

\item{inset_height}{Height of the inset as a proportion of the plot height.
Defaults to \code{0.35}.}

\item{ylim}{Y axis limits for the main plot. Passed to \code{\link[=plot_chroms]{plot_chroms()}} or
\code{\link[=plot_spectrum]{plot_spectrum()}}.}

\item{xlim}{X axis limits for the main plot. Passed to \code{\link[=plot_chroms]{plot_chroms()}} or
\code{\link[=plot_spectrum]{plot_spectrum()}}.}

\item{engine}{Plotting engine. Currently only \code{"ggplot"} is supported.}

\item{scale_spectrum}{Logical. If \code{TRUE}, scales spectrum to unit height.
Defaults to \code{FALSE}.}

\item{verbose}{Logical. If \code{TRUE} (default), prints verbose output to console.}

\item{inset_text_size}{Base size for text elements in inset. Defaults to \code{8}.}

\item{...}{Additional arguments passed to \code{\link[=plot_chroms]{plot_chroms()}} or \code{\link[=plot_spectrum]{plot_spectrum()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots a chromatogram (either a single trace or multiple traces) with an
inset UV spectrum for a specified peak. Dotted lines connect the peak apex
to the corners of the inset box.
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  data(pk_tab)
  data(Sa_warp)
  # single trace with inset spectrum
  plot_spectrum_inset("V15", peak_table = pk_tab)

  # multiple traces, inset in the center
  plot_spectrum_inset("V15", peak_table = pk_tab, idx=c(1:4),
  position = "top")

  # manual inset placement
  plot_spectrum_inset("V15", peak_table = pk_tab,
                      position = c(0.6, 0.3),
                      inset_width = 0.35, inset_height = 0.25)

  # multiple peaks with a list of positions
  plot_spectrum_inset(c("V15", "V26"), peak_table = pk_tab,
                      position = list(c(0.3, 0.5), c(0.65, 0.3)),
                      inset_width = 0.3, inset_height = 0.3)#'
                      }
}
\seealso{
\code{\link[=plot_spectrum]{plot_spectrum()}}, \code{\link[=plot_chroms]{plot_chroms()}}
}
