% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_peaks.R
\name{predict.ptw}
\alias{predict.ptw}
\title{Predict PTW}
\usage{
\method{predict}{ptw}(object, newdata, what = c("response", "time"), RTref = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{ptw}.}

\item{newdata}{an optional matrix of new data to predict. If missing,
returns the warped sample.}

\item{what}{character string specifying whether to return the warped
\code{"response"} or corrected \code{"time"}. Defaults to \code{"response"}.}

\item{RTref}{optional numeric vector of reference retention times.}

\item{...}{additional arguments (currently ignored).}
}
\value{
a matrix of warped responses or corrected retention times.
}
\description{
Predict method for objects of class \code{ptw}. Reproduced from the \code{\link[ptw:ptw]{ptw}}
package as the original is not exported.
}
\author{
Ron Wehrens
}
\keyword{internal}
