% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_chroms.R
\name{read_chroms}
\alias{read_chroms}
\title{Read chromatograms.}
\usage{
read_chroms(
  paths,
  find_files,
  format_in = c("agilent_d", "agilent_dx", "chemstation", "chemstation_fid",
    "chemstation_ch", "chemstation_csv", "chemstation_uv", "masshunter_dad",
    "chromeleon_uv", "shimadzu_ascii", "shimadzu_fid", "shimadzu_dad", "shimadzu_lcd",
    "thermoraw", "mzml", "mzxml", "waters_arw", "waters_raw", "msd", "csd", "wsd", "mdf",
    "other"),
  pattern = NULL,
  parser = c("", "chromconverter", "aston", "entab", "thermoraw", "openchrom", "rainbow"),
  format_out = c("matrix", "data.frame"),
  data_format = c("wide", "long"),
  export = FALSE,
  path_out = NULL,
  export_format = c("csv", "chemstation_csv", "cdf", "mzml", "animl"),
  read_metadata = TRUE,
  metadata_format = c("chromconverter", "raw"),
  progress_bar,
  cl = 1,
  verbose = getOption("verbose"),
  sample_names = NULL,
  dat = NULL,
  ...
)
}
\description{
Read chromatograms.
}
