% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_chroms.R
\name{reshape_chrom}
\alias{reshape_chrom}
\title{Reshapes  a single chromatogram from wide to long format}
\usage{
reshape_chrom(x)
}
\arguments{
\item{x}{A chromatographic matrix in wide format.}
}
\value{
A chromatographic matrix in long format.
}
\description{
Reshapes  a single chromatogram from wide to long format
}
\author{
Ethan Bass
}
