% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_table-methods.R
\name{row.names.peak_table}
\alias{row.names.peak_table}
\title{Return row names from a \code{peak_table} object.}
\usage{
\method{row.names}{peak_table}(x, ...)
}
\arguments{
\item{x}{A \code{\link{peak_table}} object.}

\item{...}{Additional arguments to \code{\link{row.names}}.}
}
\value{
Returns the row names of \code{peak_table$tab}.
}
\description{
These will be the names of the samples.
}
\keyword{internal}
