% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatographR-package.R
\docType{data}
\name{Sa_pr}
\alias{Sa_pr}
\title{Preprocessed goldenrod root chromatograms}
\format{
A list of four pre-processed matrices (434 retention times x 60 wavelengths).
}
\usage{
data(Sa_pr)
}
\description{
A list of four pre-processed HPLC-DAD chromatograms derived from the raw data
stored in \code{\link{Sa}}. Retention times are stored in rows and wavelengths
are stored in columns. The time axis is compressed to save space and
processing time so the data are a little choppy.
}
\seealso{
Other data objects:
\code{\link{Sa}},
\code{\link{Sa_warp}},
\code{\link{pk_tab}}
}
\concept{data objects}
\keyword{data}
