% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatographR-package.R
\docType{data}
\name{Sa_short}
\alias{Sa_short}
\title{HPLC-DAD data of goldenrod root extracts.}
\format{
Two small matrices (time x wavelength) for use in examples.
}
\description{
Subset of two HPLC-DAD data matrices of *Solidago altissima* roots samples.
}
