% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatographR-package.R
\docType{data}
\name{Sa_short_pr}
\alias{Sa_short_pr}
\title{HPLC-DAD data of goldenrod root extracts.}
\format{
Two small (pre-processed) matrices (time x wavelength) to use in
examples.
}
\description{
Subset of two HPLC-DAD data matrices of *Solidago altissima* roots samples
with pre-processing.
}
