% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatographR-package.R
\docType{data}
\name{Sa_warp}
\alias{Sa_warp}
\title{Warped goldenrod root chromatograms.}
\format{
A list of four pre-processed and warped matrices (434 times x
60 wavelengths).
}
\usage{
data(Sa_warp)
}
\description{
A list of four pre-processed and warped goldenrod root chromatograms derived
from the raw data stored in \code{\link{Sa}}.
}
\seealso{
Other data objects:
\code{\link{Sa}},
\code{\link{Sa_pr}},
\code{\link{pk_tab}}
}
\concept{data objects}
\keyword{data}
