% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_peaks.R
\name{savgol}
\alias{savgol}
\title{Savitsky Golay Smoothing from pracma}
\usage{
savgol(T, fl, forder = 4, dorder = 0)
}
\arguments{
\item{T}{Vector of signals to be filtered}

\item{fl}{Filter length (for instance fl = 51..151), has to be odd.}

\item{forder}{filter order Filter order (2 = quadratic filter, 4 = quartic).}

\item{dorder}{Derivative order (0 = smoothing, 1 = first derivative, etc.).}
}
\description{
Savitsky Golay Smoothing from pracma
}
\author{
Hans W. Borchers
}
