% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{scan_chrom}
\alias{scan_chrom}
\title{Plot spectra by clicking on the chromatogram.}
\usage{
scan_chrom(
  chrom_list,
  idx,
  lambda,
  plot_spectrum = TRUE,
  peak_table = NULL,
  scale_spectrum = FALSE,
  spectrum_labels = TRUE,
  export_spectrum = FALSE,
  ...
)
}
\arguments{
\item{chrom_list}{A list of chromatograms in matrix format (timepoints x
wavelengths). If no argument is provided here, the function will try to find
the \code{chrom_list} object using the pointer in the \code{peak_table}.}

\item{idx}{Numerical index of chromatogram you wish to plot, or "max" to
automatically select the chromatogram with the highest signal intensity at
the specified peak or retention time.}

\item{lambda}{The wavelength used for plotting chromatographic traces and
determining signal intensity.}

\item{plot_spectrum}{Logical. If \code{TRUE}, plots the spectrum of the chosen
peak. Defaults to \code{TRUE}.}

\item{peak_table}{A \code{peak_table} object created by \code{\link{get_peaktable}}.}

\item{scale_spectrum}{Logical. If \code{TRUE}, scales spectrum to unit height.
Defaults to \code{FALSE}.}

\item{spectrum_labels}{Logical. If \code{TRUE}, plots labels on maxima in spectral
plot. Defaults to \code{TRUE}.}

\item{export_spectrum}{Logical. If \code{TRUE}, exports spectrum to console.
Defaults to \code{FALSE}.}

\item{...}{Additional arguments to \code{plot_spectrum}.}
}
\value{
Invisibly returns the numeric index of the scan selected by the user.
}
\description{
Plots a chromatographic trace from the specified chromatogram (\code{idx}),
at the specified wavelength (\code{lambda}) with a dotted red line to indicate
the user-selected retention time. The trace is a single column from the
chromatographic matrix.
}
\details{
If \code{plot_spectrum} is \code{TRUE}, plots the spectrum for the specified
chromatogram at the user-specified retention time. The spectrum is a single
row from the chromatographic matrix.
}
\section{Side effects}{

If \code{engine == "base"}, plots are rendered to the active graphics device.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
data(Sa_pr)
scan_chrom(Sa_pr, lambda = "210", idx = 2, export_spectrum = TRUE)
\dontshow{\}) # examplesIf}
}
\seealso{
Other visualization functions:
\code{\link[=boxplot.peak_table]{boxplot.peak_table()}},
\code{\link[=mirror_plot]{mirror_plot()}},
\code{\link[=plot.peak_list]{plot.peak_list()}},
\code{\link[=plot.peak_table]{plot.peak_table()}},
\code{\link[=plot_all_spectra]{plot_all_spectra()}},
\code{\link[=plot_chroms]{plot_chroms()}},
\code{\link[=plot_chroms_heatmap]{plot_chroms_heatmap()}},
\code{\link[=plot_spectrum]{plot_spectrum()}}
}
\author{
Ethan Bass
}
\concept{visualization functions}
