% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks.R
\name{[.peak_list}
\alias{[.peak_list}
\title{Subset peak list}
\usage{
\method{[}{peak_list}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A \code{peak_list} object.}

\item{i}{Indices specifying elements to extract.}

\item{j}{Not used.}

\item{...}{Additional arguments (ignored).}

\item{drop}{Not used.}
}
\value{
A \code{peak_list} object.
}
\description{
Subsets a \code{peak_list} while preserving metadata attributes.
}
\author{
Ethan Bass
}
