% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_metadata.R
\name{attach_metadata}
\alias{attach_metadata}
\title{Attaches sample metadata to a \code{peak_table} object by matching sample names.}
\usage{
attach_metadata(peak_table, metadata, column)
}
\arguments{
\item{peak_table}{A \code{peak_table} object created by \code{\link{get_peaktable}}.}

\item{metadata}{A \code{data.frame} of sample metadata.}

\item{column}{Name of the column in \code{metadata} containing sample identifiers.
Must match the row names of \code{peak_table$tab}.}
}
\value{
A \code{peak_table} object with metadata stored in the \code{sample_meta} slot.
}
\description{
Metadata is provided as a \code{data.frame}, with one column containing sample
identifiers matching the row names of \code{peak_table$tab}.
}
\examples{
data(pk_tab)
path <- system.file("extdata", "Sa_metadata.csv", package = "chromatographR")
meta <- read.csv(path)
pk_tab <- attach_metadata(peak_table = pk_tab, metadata = meta, column="vial")
}
\seealso{
\code{\link{get_peaktable}}, \code{\link{normalize_data}}
}
\author{
Ethan Bass
}
