% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_metadata.R
\name{attach_metadata}
\alias{attach_metadata}
\title{Attach experimental metadata}
\usage{
attach_metadata(peak_table, metadata, column)
}
\arguments{
\item{peak_table}{Peak table}

\item{metadata}{Dataframe}

\item{column}{Name of the column with sample names.}
}
\value{
Peak table with meta-data.
}
\description{
Attach experimental meta-data to peak table.
}
\seealso{

}
\author{
Ethan Bass
}
