% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_metadata.R
\name{attach_metadata}
\alias{attach_metadata}
\title{Attach experimental metadata}
\usage{
attach_metadata(peak_table, metadata, column)
}
\arguments{
\item{peak_table}{Peak table}

\item{metadata}{Dataframe}

\item{column}{Name of the column with sample names.}
}
\value{
Peak table with meta-data.
}
\description{
Attach experimental meta-data to peak table.
}
\examples{
\dontrun{
meta <- read.csv("meta.csv")
pk_tab <- attach_metadata(peak_table = pk_tab, metadata = meta,
column="vial")}
}
\seealso{
\code{\link{get_peaktable}}
}
\author{
Ethan Bass
}
