% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_metadata.R
\name{attach_ref_spectra}
\alias{attach_ref_spectra}
\title{Attach reference spectra to a \code{peak_table} object.}
\usage{
attach_ref_spectra(peak_table, chrom_list, ref = c("max.cor", "max.int"))
}
\arguments{
\item{peak_table}{A \code{peak_table} object created by \code{\link{get_peaktable}}.}

\item{chrom_list}{Optional list of chromatograms (timepoints × wavelengths).
If \code{NULL}, the function attempts to retrieve the original \code{chrom_list}
used to generate \code{peak_table}.}

\item{ref}{Method for selecting reference spectra. Options are \code{"max.int"}
(maximum intensity) or \code{"max.cor"} (maximum average correlation).}
}
\value{
A \code{peak_table} object with reference spectra stored in the
\verb{$ref_spectra} slot.
}
\description{
Reference spectra are selected either as the spectrum with maximum intensity
(\code{"max.int"}) or as the spectrum with the highest average correlation to all
other spectra in the peak table (\code{"max.cor"}).
}
\examples{
data(pk_tab)
data(Sa_warp)
pk_tab <- attach_ref_spectra(pk_tab, ref = "max.int")
pk_tab <- attach_ref_spectra(pk_tab, ref = "max.cor")
}
\seealso{
\code{\link{get_peaks}}, \code{\link{get_peaktable}}
}
\author{
Ethan Bass
}
