% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_metadata.R
\name{attach_ref_spectra}
\alias{attach_ref_spectra}
\title{Attach reference spectra}
\usage{
attach_ref_spectra(
  peak_table,
  chrom_list = NULL,
  ref = c("max.cor", "max.int")
)
}
\arguments{
\item{peak_table}{Peak table from \code{\link{get_peaktable}}.}

\item{chrom_list}{A list of chromatograms in matrix form (timepoints x
wavelengths).}

\item{ref}{What criterion to use to select reference spectra.
Current options are maximum correlation ("max.cor") or maximum signal
intensity ("max.int").}
}
\value{
peak_table object with reference spectra attached
}
\description{
Attach reference spectra
}
\examples{

ref_m <- attach_ref_spectra(pk_tab, ref = "max.int")
}
\seealso{
\code{\link{get_peaks}}

\code{\link{get_peaktable}}
}
\author{
Ethan Bass
}
