% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.peak_table.R
\name{boxplot.peak_table}
\alias{boxplot.peak_table}
\title{Make boxplot of variables from peak table.}
\usage{
\method{boxplot}{peak_table}(x, formula, ...)
}
\arguments{
\item{x}{A peak_table object}

\item{formula}{A formula object}

\item{...}{Additional arguments to \code{\link[graphics]{boxplot}}}
}
\description{
The function can take multiple response variables on the left hand side of the
formula (separated by \code{+}). In this case, a separate boxplot will be
produced for each response variable.
}
\examples{
data(pk_tab)
path <- system.file("extdata", "Sa_metadata.csv", package = "chromatographR")
meta <- read.csv(path)
pk_tab <- attach_metadata(peak_table = pk_tab, metadata = meta, column="vial")
boxplot(pk_tab, formula=V11 ~ trt)
}
