\name{chromatographR-package}
\alias{chromatographR-package}
\alias{chromatographR}
\docType{package}
\title{chromatographR}
\description{chromatographR
}
\details{
\tabular{ll}{
Package: \tab chromatographR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2021-09-30\cr
License: GPL (>= 2)
}


}
\author{
Ethan Bass

Maintainer: Ethan Bass
}

\keyword{ package }
