% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatographR-package.R
\docType{package}
\name{chromatographR-package}
\alias{chromatographR-package}
\alias{chromatographR}
\title{chromatographR}
\description{
Chromatographic Data Analysis Toolset
}
\details{
Tools for high-throughput analysis of HPLC-DAD/UV
chromatograms (or similar data). Includes functions for preprocessing, alignment,
peak-finding and fitting, peak-table construction, data-visualization, etc.
Preprocessing and peak-table construction follow the rough formula laid out
in alsace (Wehrens, R., Bloemberg, T.G., and Eilers P.H.C., 2015.
<doi:10.1093/bioinformatics/btv299>. Alignment of chromatograms is available
using parametric time warping (ptw) (Wehrens, R., Bloemberg, T.G., and Eilers
P.H.C. 2015. <doi:10.1093/bioinformatics/btv299>) or variable penalty dynamic
time warping (VPdtw) (Clifford, D., & Stone, G. 2012. <doi:10.18637/jss.v047.i08>).
Peak-finding uses the algorithm by Tom O'Haver
<https://terpconnect.umd.edu/~toh/spectrum/PeakFindingandMeasurement.htm>.
Peaks are then fitted to a gaussian or exponential-gaussian hybrid peak shape
using non-linear least squares (Lan, K. & Jorgenson, J. W. 2001.
<doi:10.1016/S0021-9673(01)00594-5>). See the vignette for more details and
suggested workflow.

\tabular{ll}{ Package: \tab chromatographR\cr Type: \tab Package\cr Version:
\tab 0.7.0 \cr Date: \tab 2023-10-23\cr License: GPL (>= 2) }
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ethanbass.github.io/chromatographR/}
  \item \url{https://github.com/ethanbass/chromatographR/}
  \item Report bugs at \url{https://github.com/ethanbass/chromatographR/issues/}
}

}
\author{
Ethan Bass
}
