% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatographR-package.R
\docType{package}
\name{chromatographR-package}
\alias{chromatographR-package}
\alias{chromatographR}
\title{chromatographR}
\description{
chromatographR
}
\details{
\tabular{ll}{ Package: \tab chromatographR\cr Type: \tab Package\cr Version:
\tab 0.4.1 \cr Date: \tab 2022-05-18\cr License: GPL (>= 2) }
}
\author{
Ethan Bass

Maintainer: Ethan Bass
}
\keyword{package}
