% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_spectra.R
\name{cluster_spectra}
\alias{cluster_spectra}
\title{Cluster spectra}
\usage{
cluster_spectra(
  peak_table,
  alpha = 0.05,
  min_size = 5,
  max_size = NULL,
  nboot = 1000,
  plot_dend = TRUE,
  plot_spectra = TRUE,
  verbose = getOption("verbose"),
  outfile = NULL,
  save = NULL,
  parallel = TRUE,
  max.only = FALSE,
  output = c("pvclust", "clusters"),
  ...
)
}
\arguments{
\item{peak_table}{A \code{peak_table} object from \code{\link{get_peaktable}}.}

\item{alpha}{Significance threshold used to select clusters based on
bootstrap AU p-values.}

\item{min_size}{Minimum number of peaks a cluster may have.}

\item{max_size}{Maximum number of peaks a cluster may have.}

\item{nboot}{Number of bootstrap replicates for \code{\link[pvclust:pvclust]{pvclust}}.}

\item{plot_dend}{Logical. If \code{TRUE}, plots dendrogram with bootstrap values.}

\item{plot_spectra}{Logical. If \code{TRUE}, plots overlapping spectra for each
cluster.}

\item{verbose}{Logical. If \code{TRUE}, prints progress report to console.}

\item{outfile}{Optional path to an \code{.RDS} file. If not \code{NULL}, the
\code{pvclust} object is saved file using \code{saveRDS()}. Existing files are
overwritten.}

\item{save}{Logical. If \code{TRUE}, saves pvclust object to current directory.}

\item{parallel}{Logical. If \code{TRUE}, use parallel processing for
\code{\link[pvclust:pvclust]{pvclust}}.}

\item{max.only}{Logical. If \code{TRUE}, returns only highest level cluster from
each nested set of clusters.}

\item{output}{What to return. Either \code{"clusters"} to return list of clusters,
\code{"pvclust"} to return pvclust object, or \code{"both"} to return both items.}

\item{...}{Additional arguments to \code{\link[pvclust:pvclust]{pvclust}}.}
}
\value{
Returns clusters and/or \code{pvclust} object according to the value
of the \code{output} argument.
\itemize{
\item If \code{output = clusters},  returns a list of S4 \code{cluster} objects.
\item If \code{output = pvclust}, returns a \code{pvclust} object.
\item If \code{output = both}, returns a nested list containing \verb{[[1]]} the \code{pvclust}
object, and \verb{[[2]]} the list of S4 \code{cluster} objects.
}

The \code{cluster} objects consist of the following components:
\itemize{
\item \code{peaks}: a character vector containing the names of all peaks contained in
the given cluster.
\item \code{pval}: a numeric scalar containing the AU bootstrap p-value for the given
cluster.
}
}
\description{
Cluster peaks by spectral similarity.
}
\details{
Before using this function, reference spectra must be attached to the
\code{peak_table} using the \code{\link{attach_ref_spectra}} function. These spectra are
then used to construct a distance matrix derived from pairwise Pearson
correlations among peaks. Hierarchical clustering with bootstrap
resampling is performed on the resulting correlation matrix to classify peaks
by their spectral similarity as implemented in \code{\link[pvclust:pvclust]{pvclust}}.
Finally, bootstrap values can be used to select clusters that exceed a
certain confidence specified by \code{alpha}.

Clusters can be filtered by the minimum and maximum size of the cluster using
the \code{min_size} and \code{max_size} arguments respectively. Users should
be aware that the clustering algorithm will often return nested clusters.
Thus, an individual peak could appear in more than one cluster. If
\code{max.only} is \code{TRUE}, only the largest cluster in a nested tree of
clusters meeting the specified confidence threshold will be returned.

It is strongly recommended to use more than 100 bootstraps if you run the
clustering algorithm on real data even though we use \code{nboot = 100} in
the example to reduce run time. The authors of \code{pvclust} suggest using at
least 10,000 bootstrap replicates.
}
\examples{
\donttest{
if (requireNamespace("pvclust", quietly = TRUE)) {
data(pk_tab)
data(Sa_warp)
pk_tab <- attach_ref_spectra(pk_tab, Sa_warp, ref = "max.int")
cl <- cluster_spectra(pk_tab, nboot = 100, max.only = FALSE, 
save = FALSE, alpha = 0.03)
}
}
}
\references{
R. Suzuki & H. Shimodaira. 2006. Pvclust: an R package for assessing
the uncertainty in hierarchical clustering. \emph{Bioinformatics},
22(12):1540-1542. \doi{10.1093/bioinformatics/btl117}.
}
\author{
Ethan Bass
}
