% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_peaks.R
\name{combine_peaks}
\alias{combine_peaks}
\title{Combine duplicate peaks}
\usage{
combine_peaks(
  peak_table,
  tol = 0.01,
  min.cor = 0.9,
  choose = c("max", "least_sparse", "lambda"),
  lambda = NULL,
  verbose = getOption("verbose")
)
}
\arguments{
\item{peak_table}{Peak table from \code{\link{get_peaktable}}.}

\item{tol}{Tolerance for matching retention times (maximum retention time
difference). Defaults to \code{0.01}.}

\item{min.cor}{Minimum spectral correlation to confirm a match. Defaults to
\code{0.9}.}

\item{choose}{Method used to select a representative peak from each group of
duplicate peaks. \code{"max"} retains the peak with the greatest total intensity,
\code{"least_sparse"} retains the peak detected in the greatest number of samples,
and \code{"lambda"} retains peaks matching the preferred wavelength(s).}

\item{verbose}{Logical. Whether to print status to the console.}
}
\value{
A peak table derived from the original, but with columns corresponding
to duplicate peaks combined according to the specified criteria.
}
\description{
Identifies groups of duplicate peaks in a peak table and retain a single
representative peak from each group. Peaks are considered duplicates when
their retention times differ by less than \code{tol} and their
Utility function to combine duplicate peaks in peak table, (i.e., peaks that
were integrated at more than one wavelength or component) based on retention
time and/or spectral similarity. Specify tolerance (\code{tol}) for
retention time matching and minimum spectral correlation (\code{min.cor}) for
a match.
}
\examples{
data(pk_tab)
data(Sa_warp)
pk_tab <- attach_ref_spectra(pk_tab)
combine_peaks(pk_tab, tol = 0.02, min.cor = 0.9)
}
\seealso{
\code{\link{get_peaks}}

Other utility functions:
\code{\link[=get_lambdas]{get_lambdas()}},
\code{\link[=get_times]{get_times()}},
\code{\link[=merge_peaks]{merge_peaks()}},
\code{\link[=reshape_chroms]{reshape_chroms()}},
\code{\link[=reshape_peaktable]{reshape_peaktable()}}
}
\author{
Ethan Bass
}
\concept{utility functions}
