% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_peaks.R
\name{combine_peaks}
\alias{combine_peaks}
\title{Combine peaks}
\usage{
combine_peaks(
  peak_table,
  tol = 0.01,
  min.cor = 0.9,
  choose = "max",
  verbose = getOption("verbose")
)
}
\arguments{
\item{peak_table}{Peak table from \code{\link{get_peaktable}}.}

\item{tol}{Tolerance for matching retention times (maximum retention time
difference). Defaults to \code{.01}.}

\item{min.cor}{Minimum spectral correlation to confirm a match. Defaults to 
\code{0.9}.}

\item{choose}{If \code{max} will retain peak with highest intensity. Otherwise,
the first column in the data.frame will be retained.}

\item{verbose}{Logical. Whether to print status to the console.}
}
\value{
A peak table similar to the input peak table, but with duplicate
columns combined according to the specified criteria.
}
\description{
Utility function to combine duplicate peaks in peak table, i.e. peaks that
were integrated at more than one wavelength or component. Specify tolerance
(\code{tol}) for retention time matching and minimum spectral correlation
(\code{min.cor}) for a match.
}
\examples{
data(pk_tab)
data(Sa_warp)
pk_tab <- attach_ref_spectra(pk_tab)
combine_peaks(pk_tab, tol = .02, min.cor = .9)
}
\seealso{
\code{\link{get_peaks}}

Other utility functions: 
\code{\link{get_lambdas}()},
\code{\link{get_times}()},
\code{\link{merge_peaks}()}
}
\author{
Ethan Bass
}
\concept{utility functions}
