% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_peaks.R
\name{correct_peaks}
\alias{correct_peaks}
\title{Correct peak positions according to a PTW warping model}
\usage{
correct_peaks(peak_list, mod_list, chrom_list, match_names = TRUE)
}
\arguments{
\item{peak_list}{A `peak_list` object created by \code{\link{get_peaks}},
containing a nested list of peak tables where the first level is the sample,
and the second level is the spectral wavelength. Every component is described
by a matrix where each row corresponds to a feature, and the columns contain 
information on that feature (e.g., retention time, peak width (FWHM), height,
area, etc.)}

\item{mod_list}{A list of \code{ptw} models.}

\item{chrom_list}{List of chromatograms from which the ptw models are derived.}

\item{match_names}{Logical. Whether to actively match the names of the 
\code{peak_list} to the list of models (\code{mod_list}). Defaults to 
\code{TRUE}.}
}
\value{
The input list of peak tables is returned with extra columns
containing the corrected retention time.
}
\description{
Corrects retention time differences using parametric time warping as 
implemented in \code{\link[ptw]{ptw}}.
}
\details{
Once an appropriate warping model has been established, corrected retention
times can be predicted for each peak. These are stored in a separate column
in the list of peak tables.
}
\note{
This function is adapted from
\href{https://github.com/rwehrens/alsace/blob/master/R/correctPeaks.R}{getPeakTable}
function in the alsace package by Ron Wehrens.
}
\seealso{
\code{\link{correct_rt}}
}
\author{
Ron Wehrens, Ethan Bass
}
