% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_rt.R
\name{correct_rt}
\alias{correct_rt}
\title{Correct retention time}
\usage{
correct_rt(
  chrom_list,
  lambdas,
  models = NULL,
  reference = "best",
  alg = c("ptw", "vpdtw"),
  what = c("corrected.values", "models"),
  init.coef = c(0, 1, 0),
  n.traces = NULL,
  fill_zeros = FALSE,
  n.zeros = 0,
  scale = FALSE,
  trwdth = 200,
  plot_it = FALSE,
  penalty = 5,
  maxshift = 50,
  verbose = getOption("verbose"),
  show_progress = NULL,
  cl = 2,
  ...
)
}
\arguments{
\item{chrom_list}{List of chromatograms in matrix format.}

\item{lambdas}{A character or numeric vector specifying the wavelengths to
use for alignment. Only one wavelength should be specified for VPdtw warping.
For one-dimensional chromatograms, this argument can be ignored.}

\item{models}{List of models to warp by. The models provided here (if any)
must match the algorithm selected in \code{alg}.}

\item{reference}{Index of the sample that is to be considered the reference
sample.}

\item{alg}{algorithm to use: parametric time warping (\code{"ptw"}) or
variable penalty dynamic time warping (\code{"vpdtw"}).}

\item{what}{What to return: either the \code{"corrected.values"} (useful for
visual inspection and downstream analysis) or the warping \code{"models"}
(for further programmatic use).}

\item{init.coef}{Starting values for the optimization.}

\item{n.traces}{Number of traces to use.}

\item{fill_zeros}{Logical. If \code{TRUE}, out-of-bounds regions produced by
warping are filled with zeros. If \code{FALSE}, these regions are returned
as \code{NA}. Defaults to \code{FALSE}.}

\item{n.zeros}{Number of zeros to add for padding chromatograms at the edges.}

\item{scale}{Logical. If \code{TRUE}, scale chromatograms before warping.}

\item{trwdth}{width of the triangle in the WCC criterion.}

\item{plot_it}{Logical. Whether to plot alignment.}

\item{penalty}{The divisor used to calculate the penalty for
\code{\link[VPdtw]{VPdtw}}. The warping penalty is calculated by dividing the
\code{\link[VPdtw]{dilation}} by this number. Thus, a higher number will
produce a lower penalty and be more permissive, while a lower number will
produce a higher penalty and allow less warping. Defaults to \code{5}.}

\item{maxshift}{Integer. Maximum allowable shift for \code{\link[VPdtw]{VPdtw}}.
Defaults to \code{50}.}

\item{verbose}{Logical. Whether to print verbose output.}

\item{show_progress}{Logical. Whether to show progress bar. Defaults to
\code{TRUE} if \code{\link[pbapply]{pbapply}} is installed. Currently works
only for \code{ptw} alignments.}

\item{cl}{Argument to \code{\link[pbapply]{pblapply}} or
\code{\link[parallel]{mclapply}}. Either an integer specifying the number of
clusters to use for parallel processing or a cluster object created by
\code{\link[parallel]{makeCluster}}. Defaults to \code{2}. On Windows systems,
integer values will be ignored.}

\item{\dots}{Optional arguments for the \code{\link[ptw:ptw]{ptw}} function.
The only argument that cannot be changed is \code{warp.type}: this is
hard-coded to \code{"global"} to permit warping on multiple wavelengths.}
}
\value{
A list of warping models or a list of warped absorbance profiles,
according to the value of the \code{what} argument.
}
\description{
Aligns chromatograms using one of two algorithms, according to the value of
\code{alg}: either parametric time warping, as implemented in
\code{\link[ptw]{ptw}}, or variable penalty dynamic time warping, as
implemented in \code{\link[VPdtw]{VPdtw}}. The \code{init.coef} and
\code{n.traces} arguments apply only to \code{ptw} warping, while
\code{penalty} and \code{maxshift} apply only to \code{vpdtw}
warping.
}
\note{
Adapted from
\href{https://github.com/rwehrens/alsace/blob/master/R/correctRT.R}{correctRT}
function in the alsace package by Ron Wehrens.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
data(Sa_pr)
warp <- correct_rt(chrom_list = Sa_pr, lambdas=210)
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item Clifford, D., Stone, G., Montoliu, I., Rezzi, S., Martin, F. P., Guy, P.,
Bruce, S., & Kochhar, S. 2009. Alignment using variable penalty dynamic time
warping. \emph{Analytical chemistry}, \bold{81(3)}:1000-1007. \doi{10.1021/ac802041e}.
\item Clifford, D., & Stone, G. 2012. Variable Penalty Dynamic Time Warping Code
for Aligning Mass Spectrometry Chromatograms in R. \emph{Journal of
Statistical Software}, \bold{47(8)}:1-17. \doi{10.18637/jss.v047.i08}.
\item Eilers, P.H.C. 2004. Parametric Time Warping.
\emph{Anal. Chem.}, \bold{76}:404-411. \doi{10.1021/ac034800e}.
\item Wehrens, R., Bloemberg, T.G., and Eilers P.H.C. 2015. Fast
parametric time warping of peak lists. \emph{Bioinformatics},
\bold{31}:3063-3065. \doi{10.1093/bioinformatics/btv299}.
\item Wehrens, R., Carvalho, E., Fraser, P.D. 2015. Metabolite profiling in
LC–DAD using multivariate curve resolution: the alsace package for R. \emph{
Metabolomics}, \bold{11}:143-154. \doi{10.1007/s11306-014-0683-5}.
}
}
\seealso{
\code{\link[ptw:ptw]{ptw}}, \code{\link{correct_peaks}},
\code{\link[VPdtw]{VPdtw}}
}
\author{
Ethan Bass
}
