% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_rt.R
\name{correct_rt}
\alias{correct_rt}
\title{Retention time correction}
\usage{
correct_rt(
  chrom_list,
  models = NULL,
  lambdas = NULL,
  reference = "best",
  what = c("models", "corrected.values"),
  init.coef = c(0, 1, 0),
  n.traces = NULL,
  n.zeros = 0,
  scale = TRUE,
  trwdth = 200,
  ...
)
}
\arguments{
\item{chrom_list}{List of matrices containing concentration profiles.}

\item{models}{List of models to warp by.}

\item{lambdas}{Select wavelengths to use by name.}

\item{reference}{Index of the sample that is to be considered the reference
sample.}

\item{what}{What to return: either the 'corrected.values' (useful for visual
inspection) or the warping 'models' (for further programmatic use).}

\item{init.coef}{Starting values for the optimisation.}

\item{n.traces}{Number of traces to use.}

\item{n.zeros}{Number of zeros to add.}

\item{scale}{Logical. If true, scale chromatograms before warping.}

\item{trwdth}{width of the triangle in the WCC criterion.}

\item{\dots}{Optional arguments for the \code{\link[ptw:ptw]{ptw}} function.
The only argument that cannot be changed is \code{warp.type}: this is always
equal to \code{"global"}.}
}
\value{
A list of `ptw` objects or a list of warped absorbance profiles.
}
\description{
Correction of retention time differences using parametric time warping.
}
\note{
Adapted from
\url{https://github.com/rwehrens/alsace/blob/master/R/correctRT.R}{correctRT}
function in the alsace package by Ron Wehrens.
}
\examples{
\dontrun{
warping.models <- correct_rt(dat.pr, what = "models", lambdas=c('210','260','360'), n.zeros = 250)
warp <- correct_rt(chrom_list=dat.pr, models=warping.models)
}
}
\references{
Eilers, P.H.C. 2004.
\href{https://pubs.acs.org/doi/10.1021/ac034800e}{Parametric Time Warping.}
\emph{Anal. Chem.} \bold{76}:404-411.
Wehrens, R., Bloemberg, T.G., and Eilers P.H.C. 2015.
\href{https://academic.oup.com/bioinformatics/article/31/18/3063/240579}{Fast
parametric time warping of peak lists.} \emph{Bioinformatics}
\bold{31}:3063-3065.
Wehrens, R., Carvalho, E., Fraser, P.D. 2015.
\href{https://doi.org/10.1007/s11306-014-0683-5}{Metabolite profiling in
LC–DAD using multivariate curve resolution: the alsace package for R.} \emph{
Metabolomics} \bold{11:1}:143-154.
}
\seealso{
\code{\link[ptw:ptw]{ptw}}, \code{\link{correct_peaks}}
}
\author{
Ethan Bass
}
