% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_rt.R
\name{correct_rt}
\alias{correct_rt}
\title{Correct retention time}
\usage{
correct_rt(
  chrom_list,
  models = NULL,
  lambdas = NULL,
  reference = "best",
  what = c("models", "corrected.values"),
  init.coef = c(0, 1, 0),
  n.traces = NULL,
  n.zeros = 0,
  scale = TRUE,
  trwdth = 200,
  ...
)
}
\arguments{
\item{chrom_list}{List of matrices containing concentration profiles.}

\item{models}{List of models to warp by.}

\item{lambdas}{Select wavelengths to use by name.}

\item{reference}{Index of the sample that is to be considered the reference
sample.}

\item{what}{What to return: either the 'corrected.values' (useful for visual
inspection) or the warping 'models' (for further programmatic use).}

\item{init.coef}{Starting values for the optimization.}

\item{n.traces}{Number of traces to use.}

\item{n.zeros}{Number of zeros to add.}

\item{scale}{Logical. If true, scale chromatograms before warping.}

\item{trwdth}{width of the triangle in the WCC criterion.}

\item{\dots}{Optional arguments for the \code{\link[ptw:ptw]{ptw}} function.
The only argument that cannot be changed is \code{warp.type}: this is always
equal to \code{"global"}.}
}
\value{
A list of `ptw` objects or a list of warped absorbance profiles.
}
\description{
Corrects retention time differences using parametric time warping as 
implemented in \code{\link[ptw:ptw]{ptw}}.
}
\note{
Adapted from
\href{https://github.com/rwehrens/alsace/blob/master/R/correctRT.R}{correctRT}
function in the alsace package by Ron Wehrens.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(Sa_pr)
warping.models <- correct_rt(Sa_pr, what = "models", lambdas=c("210"))
warp <- correct_rt(chrom_list = Sa_pr, models = warping.models)
\dontshow{\}) # examplesIf}
}
\references{
Eilers, P.H.C. 2004. Parametric Time Warping.
\emph{Anal. Chem.} \bold{76}:404-411. \doi{10.1021/ac034800e}.

Wehrens, R., Bloemberg, T.G., and Eilers P.H.C. 2015. Fast
parametric time warping of peak lists. \emph{Bioinformatics}
\bold{31}:3063-3065. \doi{10.1093/bioinformatics/btv299}.

Wehrens, R., Carvalho, E., Fraser, P.D. 2015.
Metabolite profiling in
LC–DAD using multivariate curve resolution: the alsace package for R. \emph{
Metabolomics} \bold{11:1}:143-154. \doi{10.1007/s11306-014-0683-5}
}
\seealso{
\code{\link[ptw:ptw]{ptw}}, \code{\link{correct_peaks}}
}
\author{
Ethan Bass
}
