% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_peaks.R
\name{filter_peaktable}
\alias{filter_peaktable}
\title{Filter peak table}
\usage{
filter_peaktable(
  peak_table,
  rts,
  min_rt,
  max_rt,
  min_value,
  max_zeros,
  lambda,
  what = c("median", "mean", "max"),
  tol = 0
)
}
\arguments{
\item{peak_table}{A peak_table object from \code{\link{get_peaktable}}.}

\item{rts}{Vector of retention times used to select features. Values are
mapped to the closest retention times in the peak table. If no
match is found within \code{tol}, the value is ignored and a warning is issued.}

\item{min_rt}{Minimum retention time for features to be retained.}

\item{max_rt}{Maximum retention time for features to be retained.}

\item{min_value}{Minimum threshold for feature intensity summarized across
samples (using the method specified by \code{what}).}

\item{max_zeros}{Maximum allowed feature sparsity (proportion of zero values
across samples).}

\item{lambda}{Wavelength(s) used to select features in the peak table.
Only features matching the specified wavelength(s) are retained.}

\item{what}{Method used to summarize feature intensity across samples for
filtering. One of \code{"mean"}, \code{"median"} (default),  or \code{"max"}.}

\item{tol}{Tolerance for matching of retention times to \code{rts}. A
feature is considered a match if the absolute difference is ≤ \code{tol}.}
}
\value{
A filtered \code{peak_table} containing only features (columns) that
satisfy the specified criteria. The same feature indices are applied to all
\code{peak_table} components.
}
\description{
Utility function to filter features (columns) in a \code{peak_table}. Filtering is
applied consistently across all components of the peak table, including the
intensity matrix (\code{tab}), feature metadata (\code{pk_meta}), and reference spectra
(if present). Filtering can be based on retention time, wavelength, feature
intensity summarized across samples (mean, median, or max), and/or feature
sparsity (i.e., proportion of zero values across samples).
}
\examples{
data(pk_tab)
pk_tab <- filter_peaktable(pk_tab, min_rt = 10, max_rt = 16)
}
\seealso{
\code{\link{get_peaktable}}, \code{\link{filter_peaks}}
}
\author{
Ethan Bass
}
