% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_purity.R, R/peak_purity.R
\name{find_noise}
\alias{find_noise}
\title{Define noise spectra based on specified threshold}
\usage{
find_noise(x, noise_threshold = 0.01, lambdas)

find_noise(x, noise_threshold = 0.01, lambdas)
}
\arguments{
\item{x}{A chromatogram in matrix format}

\item{noise_threshold}{Threshold to define noise. Highest proportion of maximum absorbance.}

\item{lambdas}{Wavelengths to include.}
}
\description{
Define noise spectra based on specified threshold

Define noise spectra based on specified threshold
}
\author{
Ethan Bass
}
