% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_peaks.R
\name{find_peaks}
\alias{find_peaks}
\title{Find peaks}
\usage{
find_peaks(
  y,
  smooth_type = c("gaussian", "box", "savgol", "mva", "tmva", "none"),
  smooth_window = 0.001,
  slope_thresh = 0,
  amp_thresh = 0,
  bounds = TRUE
)
}
\arguments{
\item{y}{A chromatographic signal (as a numeric vector).}

\item{smooth_type}{Type of smoothing. Either gaussian kernel (\code{"gaussian"}),
box kernel (\code{"box"}), Savitzky-Golay smoothing (\code{"savgol"}),
moving average (\code{"mva"}), triangular moving average (\code{"tmva"}), or
no smoothing (\code{"none"}).}

\item{smooth_window}{Smoothing window. Larger values of this parameter will
exclude sharp, narrow features. If the supplied value is between 0 and
1, the window will be interpreted as a proportion of points to include.
Otherwise, the window will be interpreted as the absolute number of points to
include in the window. Defaults to \code{0.001}.}

\item{slope_thresh}{Minimum threshold for slope of the smoothed first
derivative. This parameter filters peaks on the basis of their width, such
that larger values will exclude broad peaks from the peak list. Defaults to
\code{0}.}

\item{amp_thresh}{Minimum threshold for peak amplitude. This parameter
filters on the basis of peak height, such that larger values will exclude
small peaks from the peak list. Defaults to \code{0}.}

\item{bounds}{Logical. If \code{TRUE} (default), includes peak boundaries in
\code{data.frame}.}
}
\value{
If \code{bounds == TRUE}, returns a \code{data.frame} containing the center,
start, and end of each identified peak. Otherwise, returns a numeric vector
of peak centers. All locations are expressed as indices.
}
\description{
Detects peaks in a chromatographic profile by identifying zero-crossings in
the smoothed first derivative of the signal (\code{y}) that exceed a specified
slope threshold (\code{slope_thresh}). Smoothing reduces spurious local extrema
that do not represent true features. Peaks can be filtered using a minimum
amplitude threshold (\code{amp_thresh}), which removes low-intensity peaks.
}
\details{
Available smoothing methods include Gaussian kernel (\code{"gaussian"}),
box kernel (\code{"box"}), Savitzky–Golay (\code{"savgol"}), moving average (\code{"mva"}),
triangular moving average (\code{"tmva"}), and no smoothing (\code{"none"}).

Preprocessing with \code{\link{preprocess}} is recommended prior to peak detection.
Overly high chromatographic resolution may sometimes cause peak splitting.
In such cases, it is recommended to either increase the \code{smooth_window} or
reduce the time-axis resolution by adjusting the \code{dim1} argument during
preprocessing.
}
\note{
The \code{find_peaks} function is adapted from MATLAB code included in
Prof. Tom O'Haver's \href{https://terpconnect.umd.edu/~toh/spectrum/PeakFindingandMeasurement.htm}{Pragmatic Introduction to Signal Processing}.
}
\examples{
data(Sa_pr)
find_peaks(Sa_pr[[1]][,"220"])
}
\references{
O'Haver, Tom. Pragmatic Introduction to Signal Processing:
Applications in scientific measurement. \url{https://terpconnect.umd.edu/~toh/spectrum/}
(Accessed January, 2022).
}
\seealso{
\code{\link{fit_peaks}}, \code{\link{get_peaks}}
}
\author{
Ethan Bass
}
\keyword{internal}
