% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_peaks.R
\name{find_peaks}
\alias{find_peaks}
\title{Find peaks in chromatographic profile}
\usage{
find_peaks(
  y,
  smooth_type = c("gaussian", "box", "savgol", "mva", "tmva", "none"),
  smooth_window = 0.001,
  slope_thresh = 0,
  amp_thresh = 0,
  bounds = TRUE
)
}
\arguments{
\item{y}{response (numerical vector)}

\item{smooth_type}{Type of smoothing. Either gaussian kernel (\code{gaussian}),
box kernel (\code{box}), savitzky-golay smoothing (\code{savgol}),
moving average (\code{mva}), triangular moving average (\code{tmva}), or 
no smoothing (\code{none}).}

\item{smooth_window}{Smoothing window. If the supplied value is between 0 and
1, window will be interpreted as a proportion of points to include. Otherwise,
the window will be the absolute number of points to include in the window.
(Defaults to .001).}

\item{slope_thresh}{Minimum threshold for peak slope. (Defaults to 0).}

\item{amp_thresh}{Minimum threshold for peak amplitude. (Defaults to 0).}

\item{bounds}{Logical. If TRUE, includes peak boundaries in data.frame.
(Defaults to TRUE).}
}
\value{
If bounds == TRUE, returns a data.frame containing the center, start,
and end of each identified peak. Otherwise, returns a numeric vector of peak
centers. All locations are expressed as indices.
}
\description{
Find peaks in chromatographic profile.
}
\details{
Find peaks by looking for zero-crossings in the smoothed first derivative of
the signal (\code{y}) that exceed the specified slope threshold
(\code{slope_thresh}). Peaks can also be filtered by supplying a minimal
amplitude threshold (\code{amp_thresh}), to filter out peaks below the
specified height. 

It is recommended to do pre-processing using the \code{\link{preprocess}}
function before peak detection. Overly high chromatographic resolution can 
sometimes lead to peaks being split into multiple segments. In this case,
it is recommended to reduce the chromatogram resolution using the \code{dim2}
argument before proceeding with further analyses.
}
\note{
The \code{find_peaks} function is adapted from matlab code in Prof.
Tom O'Haver's
\href{http://terpconnect.umd.edu/~toh/spectrum/PeakFindingandMeasurement.htm}{
Pragmatic Introduction to Signal Processing}.
}
\examples{
data(Sa_pr)
find_peaks(Sa_pr[[1]][,"220"])
}
\references{
O'Haver, Tom. Pragmatic Introduction to Signal Processing:
Applications in scientific measurement.
/href{https://terpconnect.umd.edu/~toh/spectrum/} (Accessed January, 2022).
}
\seealso{
\code{\link{fit_peaks}}, \code{\link{get_peaks}}
}
\author{
Ethan Bass
}
