% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_peaks.R
\name{fit_peaks}
\alias{fit_peaks}
\title{Fit chromatographic peaks to an exponential-gaussian hybrid or gaussian
profile}
\usage{
fit_peaks(
  x,
  lambda,
  pos = NULL,
  sd_max = 50,
  fit = c("egh", "gaussian", "raw"),
  max_iter = 1000,
  estimate_purity = TRUE,
  noise_threshold = 0.001,
  ...
)
}
\arguments{
\item{x}{A single chromatogram in matrix format.}

\item{lambda}{Wavelength to fit peaks at.}

\item{pos}{Locations of peaks in vector \code{y}. If NULL, \code{find_peaks} will
run automatically to find peak positions.}

\item{sd_max}{Maximum width (standard deviation) for peaks. Defaults to 50.}

\item{fit}{Function for peak fitting. Currently, exponential-gaussian hybrid
\code{egh}, \code{gaussian} and \code{raw} settings are supported. If \code{
raw} is selected, trapezoidal integration will be performed on raw data
without fitting a peak shape. Defaults to \code{egh}.)}

\item{max_iter}{Maximum number of iterations to use in nonlinear least
squares peak-fitting. Defaults to 1000.}

\item{estimate_purity}{Logical. Whether to estimate purity or not. Defaults
to TRUE.}

\item{noise_threshold}{Noise threshold. Input to \code{get_purity}.}

\item{...}{Additional arguments to \code{find_peaks}.}
}
\value{
The \code{fit_peaks} function returns a matrix, whose columns contain
the following information about each peak:
\item{rt}{Location of the peak maximum.}
\item{start}{Start of peak (only included in table if \code{bounds = TRUE}).}
\item{end}{End of peak (only included in table if \code{bounds = TRUE}).}
\item{sd}{The standard deviation of the peak.}
\item{tau}{\eqn{\tau} parameter (only included in table if \code{fit = "egh"}).}
\item{FWHM}{The full width at half maximum.}
\item{height}{Peak height.}
\item{area}{Peak area.}
\item{r.squared}{The R\if{html}{\out{<sup>}}2\if{html}{\out{</sup>}} value for linear fit of the model to the data.}
\item{purity}{The spectral purity of peak as assessed by \code{\link{get_purity}}.}
Again, the first five elements (rt, start, end, sd and FWHM) are expressed
as indices, so not in terms of the real retention times. The transformation
to "real" time is done in function \code{get_peaks}.
}
\description{
Fit peak parameters using exponential-gaussian hybrid or gaussian function.
}
\details{
Peak parameters are calculated by fitting the data
to a gaussian or exponential-gaussian hybrid curve using non-linear least
squares estimation as implemented in \code{\link[minpack.lm:nlsLM]{nlsLM}}.
The area under the fitted curve is then estimated using trapezoidal
approximation.
}
\note{
The \code{\link{fit_peaks}} function is adapted from Dr. Robert
Morrison's
\href{https://github.com/robertdouglasmorrison/DuffyTools}{DuffyTools package}
as well as code published in Ron Wehrens'
\href{https://github.com/rwehrens/alsace}{alsace} package.
}
\examples{
data(Sa_pr)
fit_peaks(Sa_pr[[1]], lambda = 220)
}
\references{
\itemize{
\item Lan, K. & Jorgenson, J. W. 2001. A hybrid of exponential and gaussian
functions as a simple model of asymmetric chromatographic peaks. \emph{Journal of
Chromatography A} \bold{915}:1-13. \doi{10.1016/S0021-9673(01)00594-5}.
\item Naish, P. J. & Hartwell, S. 1988. Exponentially Modified Gaussian functions - A
good model for chromatographic peaks in isocratic HPLC? \emph{Chromatographia},
\bold{26}: 285-296. \doi{10.1007/BF02268168}.
}
}
\seealso{
\code{\link{find_peaks}}, \code{\link{get_peaks}}
}
\author{
Ethan Bass
}
\keyword{internal}
