% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_purity.R
\name{get_agilent_threshold}
\alias{get_agilent_threshold}
\title{Calculate purity thresholds}
\usage{
get_agilent_threshold(
  x,
  pos,
  weight = 1,
  noise_variance = NULL,
  noise_threshold = 0.005,
  lambdas
)
}
\arguments{
\item{x}{A chromatogram in matrix format.}

\item{pos}{A vector containing peak information.}

\item{weight}{Scaling parameter affecting stringency of threshold. Defaults
to \code{1}.}

\item{noise_variance}{Variance of noise.}

\item{noise_threshold}{Threshold to define noise. Highest proportion of
maximum absorbance. Defaults to \code{0.005}.}

\item{lambdas}{Wavelengths to include.}
}
\value{
Returns a vector of purity thresholds at each retention time index
within the peak specified by \code{pos}.
}
\description{
Intermediate function used in \code{\link{get_purity}} to compute variance-based
thresholds.
}
\references{
Stahl, Mark. “Peak Purity Analysis in HPLC and CE Using Diode-Array Technology.”
Agilent Technologies, April 1, 2003, 16.
\href{https://web.archive.org/web/20220615145615/https://www.agilent.com/cs/library/applications/5988-8647EN.pdf}{5988-8647EN.pdf}
}
\author{
Ethan Bass
}
\keyword{internal}
