% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_purity.R, R/peak_purity.R
\name{get_agilent_threshold}
\alias{get_agilent_threshold}
\title{calculate purity thresholds}
\usage{
get_agilent_threshold(
  x,
  pos,
  weight = 1,
  noise_threshold = 0.005,
  lambdas = c(210:400)
)

get_agilent_threshold(
  x,
  pos,
  weight = 1,
  noise_threshold = 0.005,
  lambdas = c(210:400)
)
}
\arguments{
\item{x}{A chromatogram in matrix format}

\item{pos}{A vector containing peak information}

\item{weight}{scaling parameter affecting stringency of threshold}

\item{noise_threshold}{Threshold to define noise. Highest proportion of maximum absorbance.}

\item{lambdas}{Wavelengths to include}
}
\description{
calculate purity thresholds

calculate purity thresholds
}
\author{
Ethan Bass
}
