% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_lambdas}
\alias{get_lambdas}
\title{Get lambdas}
\usage{
get_lambdas(x)
}
\arguments{
\item{x}{List of chromatograms or \code{peak_table} object.}
}
\value{
Numeric vector of wavelengths.
}
\description{
Get wavelengths from a list of chromatograms or a peak_table object.
}
\seealso{
Other utility functions: 
\code{\link{combine_peaks}()},
\code{\link{get_times}()},
\code{\link{merge_peaks}()},
\code{\link{reshape_chroms}()},
\code{\link{reshape_peaktable}()}
}
\concept{utility functions}
