% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks.R
\name{get_peaks}
\alias{get_peaks}
\title{Extract all peaks from a list of chromatograms.}
\usage{
get_peaks(
  chrom_list,
  lambdas,
  fit = c("egh", "gaussian"),
  sd.max = 50,
  max.iter = 100,
  ...
)
}
\arguments{
\item{chrom_list}{A list of profile matrices, each of the same dimensions
(timepoints x wavelengths).}

\item{lambdas}{Character vector of wavelengths to find peaks at.}

\item{fit}{What type of fit to use. Current options are gaussian and
exponential gaussian hybrid.}

\item{sd.max}{Maximum width (standard deviation) for peaks. Defaults to 50.}

\item{max.iter}{Maximum number of iterations for non-linear least squares
in \code{\link{fit_peaks}}.}

\item{\dots}{Additional arguments to \code{\link{find_peaks}}.}
}
\value{
The result is a list, with each element corresponding to one data
file, and containing data for the fitted peaks for each of the ALS
components. Note that this function presents the "rt", "sd" and "FWHM"
fields in real time units.
}
\description{
Extracts all peaks in a list of chromatograms.
}
\details{
Peaks are located by finding zero-crossings in the smoothed first derivative
of the specified chromatographic traces (function \code{\link{find_peaks}}).
At the given positions, an exponential-gaussian hybrid (or regular gaussian)
function is fit to the signal using \code{\link{fit_peaks}}). The area is then
calculated using a trapezoidal approximation.
}
\note{
Function is adapted from the
\url{https://github.com/rwehrens/alsace/blob/master/R/getAllPeaks.R}{getAllPeaks}
function authored by Ron Wehrens.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(Sa_short_pr)
pks <- get_peaks(Sa_short_pr, lambdas = c('210'), sd.max=50, fit="egh")
\dontshow{\}) # examplesIf}
}
\references{
Wehrens, R., Carvalho, E., Fraser, P.D. 2015.
\href{https://doi.org/10.1007/s11306-014-0683-5}{Metabolite profiling in
LC–DAD using multivariate curve resolution: the alsace package for R.} \emph{
Metabolomics} \bold{11:1}:143-154.
}
\author{
Ethan Bass
}
