% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaktable.R
\name{get_peaktable}
\alias{get_peaktable}
\title{Convert peak lists into an aligned peak table.}
\usage{
get_peaktable(
  peak_list,
  chrom_list,
  response = c("area", "height"),
  use.cor = NULL,
  hmax = 0.2,
  summarize_by = c("weighted.mean", "median", "mean", "max"),
  plot_it = FALSE,
  ask = plot_it,
  clust = c("rt", "sp.rt"),
  sigma.t = NULL,
  sigma.r = 0.5,
  deepSplit = FALSE,
  verbose = FALSE,
  out = c("data.frame", "matrix")
)
}
\arguments{
\item{peak_list}{A \code{peak_list} object created by \code{\link{get_peaks}}, containing a
nested list of features, where the first level is the sample, and the second
level is the spectral wavelength. Every sample x wavelength component is
described by a  \code{data.frame} with a row for each peak and columns containing
information on various peak parameters.}

\item{chrom_list}{A list of chromatographic matrices. Only required when
\code{clust = "sp.rt"}.}

\item{response}{Peak intensity measure: either \code{"area"} (default) or
\code{"height"}.}

\item{use.cor}{Logical. Indicates whether to use corrected retention times
(\code{rt.cor} column) or raw retention times \code{rt} column). Unless
otherwise specified, the \code{rt.cor} column will be used by default if it
exists in the provided \code{peak_list}.}

\item{hmax}{Height at which the complete linkage dendrogram will be cut. Can
be interpreted as the maximal intercluster retention time difference.}

\item{summarize_by}{How to select the representative peak from each cluster.
Options are \code{"weighted.mean"} (default, weights peaks by their intensity as
given by \code{response}), \code{"mean"}, \code{"median"} (which aggregate metadata across
all peaks in the cluster) or \code{"max"} (which selects the most intense peak in
the cluster and uses its metadata directly).}

\item{plot_it}{Logical. If \code{TRUE}, for every component a strip plot will be
shown indicating the clustering.}

\item{ask}{Logical. Ask before showing new plot? Defaults to \code{TRUE}.}

\item{clust}{Specify whether to perform hierarchical clustering based on
spectral similarity and retention time (\code{sp.rt}) or retention time alone
(\code{rt}). Defaults to \code{rt}.}

\item{sigma.t}{Width of Gaussian kernel for retention time similarity.
Controls weight given to retention time when \code{"sp.rt"} is selected.}

\item{sigma.r}{Width of Gaussian kernel for spectral similarity. Controls
weight given to spectral correlation if \code{"sp.rt"} is selected.}

\item{deepSplit}{Logical. Controls sensitivity to cluster splitting. If
\code{TRUE}, function will return a larger number of smaller clusters. For
additional information, see \code{\link[dynamicTreeCut:cutreeDynamic]{cutreeDynamic}}.}

\item{verbose}{Logical. Whether to print warning when combining peaks into
single time window. Defaults to \code{FALSE}.}

\item{out}{Specify \code{data.frame} (default) or \code{matrix} as output.}
}
\value{
The function returns an S3 \code{\link{peak_table}} object, containing the
following elements:
\itemize{
\item \code{tab}: The peak table itself -- a \code{data.frame} of intensities in a
sample × peak configuration.
\item \code{pk_meta}: A \code{data.frame} containing peak metadata (e.g., the spectral
component, peak number, and average retention time), summarized across all
peaks in each cluster according to \code{summarize_by}.
\item \code{sample_meta}: A \code{data.frame} of sample metadata. Must be added using
\code{\link{attach_metadata}}.
\item \code{ref_spectra}: A \code{data.frame} of reference spectra (in a wavelength ×
peak configuration). Must be added using \code{\link{attach_ref_spectra}}.
\item \code{args}: A vector of arguments given to \code{\link{get_peaktable}} to
generate the peak table.
}
}
\description{
Converts a \code{peak_list} object containing peaks from multiple samples into
an aligned peak table by clustering peaks across samples according to
retention time and (optionally) spectral similarity. Clusters are defined
by cutting a complete-linkage dendrogram at height \code{hmax}, which can be
understood as the maximal allowed clustering distance.
}
\details{
By default, clustering is performed on retention times alone (when
\code{clust = "rt"}). Alternatively, clustering can also incorporate information
about spectral similarity (when \code{clust = "sp.rt"}) using a similarity function
adapted from Broeckling et al., 2014:
\deqn{
S_{ij} =
  \exp\left(-\frac{(t_i - t_j)^2}{2\sigma_t^2}\right)
   \cdot
   \exp\left(-\frac{(1 - c_{ij})^2}{2\sigma_r^2}\right)
}
where \eqn{c_{ij}} is the spectral correlation between spectra \eqn{i} and
\eqn{j}, and \eqn{\sigma_t} and \eqn{\sigma_r} control the relative
contributions of retention time and spectral similarity, respectively (see
\code{sigma.t} and \code{sigma.r}).

If two peaks from the same sample are assigned to the same cluster, a warning
message is printed to the console. These warnings are often benign but may
also indicate that \code{hmax} is too large. Reducing \code{hmax} may prevent excessive
clustering, although overly small values may instead lead to splitting of
peaks across multiple clusters. Filtering low-intensity peaks prior to
clustering may also help reduce these warnings.

Once clusters are formed, peak metadata are summarized within each cluster
according to \code{summarize_by}. By default (\code{"weighted.mean"}), peak metadata
are averaged using peak intensity as weights, reducing the influence of small
peaks which are more likely to represent noise. Alternatively, the \code{"max"}
argument selects the metadata associated with the most intense peak in each
cluster.
}
\note{
This function is adapted from the \code{getPeakTable} function in the alsace
package by Ron Wehrens: \url{https://github.com/rwehrens/alsace/blob/master/R/getPeakTable.R}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
data(Sa_pr)
pks <- get_peaks(Sa_pr, lambdas = c('210'))
get_peaktable(pks, response = "area")
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item Broeckling, C. D., Afsar F.A., Neumann S., Ben-Hur A., and Prenni J.E. 2014.
RAMClust: A Novel Feature Clustering Method Enables Spectral-Matching-Based
Annotation for Metabolomics Data. \emph{Anal. Chem.} 86:6812-6817.
\doi{10.1021/ac501530d}.
\item Wehrens, R., Carvalho, E., Fraser, P.D. 2015. Metabolite profiling in
LC–DAD using multivariate curve resolution: the alsace package for R.
\emph{Metabolomics} 11:143-154. \doi{10.1007/s11306-014-0683-5}.
}
}
\seealso{
\code{\link[=peak_table-class]{peak_table}}, \code{\link{attach_ref_spectra}}, \code{\link{attach_metadata}}
}
\author{
Ethan Bass
}
