% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_purity.R
\name{get_purity}
\alias{get_purity}
\title{Calculate mean peak purity}
\usage{
get_purity(x, pos, weight = 1, cutoff = 0.05)
}
\arguments{
\item{x}{A chromatogram in matrix format}

\item{pos}{A vector containing peak information}

\item{weight}{weight provided to \code{\link{get_agilent_threshold}}}

\item{cutoff}{Proportion of maximum absorbance to use as cutoff.
Argument to \code{\link{trim_peak}}.}
}
\value{
Returns the mean purity of the peak specified by \code{pos}, defined
as the proportion of timepoints with purity values below 1.
}
\description{
Estimates peak purity by assessing the dissimilarity of the spectra
comprising the peak, using the method described by Stahl.
}
\references{
Stahl, Mark. “Peak Purity Analysis in HPLC and CE Using Diode-Array Technology.”
Agilent Technologies, April 1, 2003, 16.
/href{https://www.agilent.com/cs/library/applications/5988-8647EN.pdf}
}
\author{
Ethan Bass
}
