% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_purity.R
\name{get_purity_values}
\alias{get_purity_values}
\title{Calculate peak purity values}
\usage{
get_purity_values(
  x,
  pos,
  weight = 1,
  noise_variance = NULL,
  noise_threshold = 0.005,
  lambdas
)
}
\arguments{
\item{x}{A chromatogram in matrix format}

\item{pos}{A vector containing peak information}

\item{weight}{weight provided to \code{\link{get_agilent_threshold}}.}

\item{noise_variance}{Variance of noise. Argument to \code{\link{get_agilent_threshold}}.}

\item{noise_threshold}{Threshold to define noise. Highest proportion of maximum absorbance.}

\item{lambdas}{Wavelengths to include in calculations.}
}
\value{
Returns a vector of peak purity values at each timepoint within the
peak specified by \code{pos}.
}
\description{
Calculate peak purity values
}
\references{
Stahl, Mark. “Peak Purity Analysis in HPLC and CE Using Diode-Array Technology.”
Agilent Technologies, April 1, 2003, 16.
/href{https://www.agilent.com/cs/library/applications/5988-8647EN.pdf}
}
\author{
Ethan Bass
}
