% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_purity.R, R/peak_purity.R
\name{get_spectral_similarity}
\alias{get_spectral_similarity}
\title{Calculate spectral similarity}
\usage{
get_spectral_similarity(x, pos)

get_spectral_similarity(x, pos)
}
\arguments{
\item{x}{A chromatogram in matrix format}

\item{pos}{A vector containing peak information}
}
\description{
Calculate spectral similarity

Calculate spectral similarity
}
\author{
Ethan Bass
}
