% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_times}
\alias{get_times}
\title{Get retention times}
\usage{
get_times(x, idx = 1)
}
\arguments{
\item{x}{A list of chromatograms or \code{peak_table} object.}

\item{idx}{Index of chromatogram from which to extract times.}
}
\value{
Numeric vector of retention times from the chromatogram specified by
\code{idx}.
}
\description{
Get retention times from a list of chromatograms or a \code{peak_table} object.
}
\examples{
data(Sa_warp)
get_times(Sa_warp)
}
\seealso{
Other utility functions:
\code{\link[=combine_peaks]{combine_peaks()}},
\code{\link[=get_lambdas]{get_lambdas()}},
\code{\link[=merge_peaks]{merge_peaks()}},
\code{\link[=reshape_chroms]{reshape_chroms()}},
\code{\link[=reshape_peaktable]{reshape_peaktable()}}
}
\concept{utility functions}
\keyword{internal}
