% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_times}
\alias{get_times}
\title{Get retention times}
\usage{
get_times(x, idx = 1)
}
\arguments{
\item{x}{List of chromatograms or \code{peak_table} object.}

\item{idx}{Index of chromatogram from which to extract times}
}
\value{
Numeric vector of retention times from the chromatogram specified by
\code{idx}.
}
\description{
Get retention times from a list of chromatograms or a peak_table object.
}
