% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_chromes.R
\name{load_chromes}
\alias{load_chromes}
\title{Import folders of chromatograms in CSV format.}
\usage{
load_chromes(paths, pattern = "CSV", sep = ",", dat = NULL, ...)
}
\arguments{
\item{paths}{Path(s) to folders where chromatograms are stored.}

\item{pattern}{Pattern to identify files (i.e. a file extension).
Defaults to 'CSV'.}

\item{sep}{Argument provided to \code{read.csv}. Defaults to ",".}

\item{dat}{Optional list of chromatograms. If list is provided, the function
will append newly imported chromatograms to the existing list.}

\item{...}{Additional arguments to \code{\link{read.csv}}.}
}
\value{
A list of chromatograms in matrix format.
}
\description{
Convenience function to import chromatograms from a list of folders.
Chromatograms must be in CSV format.
}
\examples{
\dontrun{
###  import from single folder
dat <- load_chromes(paths = path)
### import from multiple folders
path = 'foo'
folders <- list.files(path = path, pattern = "EXPORT3D")
dat <- load_chromes(folders)
}
}
\author{
Ethan Bass
}
