% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_peaks.R
\name{merge_peaks}
\alias{merge_peaks}
\title{Merge split peaks in peak table}
\usage{
merge_peaks(peak_table, peaks)
}
\arguments{
\item{peak_table}{Peak table from \code{\link{get_peaktable}}.}

\item{peaks}{A vector specifying the names or indices of peaks to be merged.}
}
\value{
A peak table similar to the input peak table, but where the specified
columns are combined combined.
}
\description{
Merges the specified peaks, by selecting the largest value from each column.
Utility function to combine split peaks into a single column of the peak table.
}
