% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize peak table or chromatograms}
\usage{
normalize_data(
  peak_table,
  column,
  chrom_list = NULL,
  what = c("peak_table", "chrom_list"),
  by = NULL,
  on_invalid = c("warn", "error", "silent")
)
}
\arguments{
\item{peak_table}{A \code{peak_table} object.}

\item{column}{The name of the column to be used for normalization.}

\item{chrom_list}{List of chromatograms for normalization. The samples must
be in same order as the \code{peak_table}. If omitted, the function will attempt to
find it automatically using the pointer from the \code{peak_table}.}

\item{what}{Output type to return: either \code{"peak_table"} (default) or
\code{"chrom_list"}.}

\item{by}{Whether to normalize by a column in sample metadata (\code{meta}) or by
a column in the peak table (\code{peak}). By default, this parameter is inferred
based on the \code{column} name.}

\item{on_invalid}{How to handle invalid normalization values (i.e. zero,
negative, or \code{NA} values). One of \code{"warn"} (the default), \code{"silent"}, or
\code{"error"}. Invalid values are replaced with \code{NA} unless \code{"error"} is chosen.}
}
\value{
Either a normalized \code{peak_table} object or a normalized \code{chrom_list},
depending on the value of \code{what}.
}
\description{
Normalizes peak table or list of chromatograms by a column in the sample
metadata or in the peak table. For normalization by sample metadata, the
metadata must first be attached to the \code{peak_table} using \code{\link{attach_metadata}}.
}
\examples{
data(pk_tab)
path <- system.file("extdata", "Sa_metadata.csv", package = "chromatographR")
meta <- read.csv(path)

# normalize by sample mass
pk_tab <- attach_metadata(peak_table = pk_tab, metadata = meta, column="vial")
norm <- normalize_data(pk_tab, "mass", what = "peak_table")

# normalize by internal standard
norm <- normalize_data(pk_tab, column = "V16", by = "peak")
}
\seealso{
\code{\link{get_peaktable}} \code{\link{attach_metadata}}
}
\author{
Ethan Bass
}
