% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_metadata.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize peak table or chromatograms}
\usage{
normalize_data(
  peak_table,
  column,
  chrom_list,
  what = c("peak_table", "chrom_list"),
  by = c("meta", "peak")
)
}
\arguments{
\item{peak_table}{A `peak_table` object}

\item{column}{The name of the column containing the weights.}

\item{chrom_list}{List of chromatograms for normalization. The samples must
be in same order as the peak_table. If no argument is provided here, the
function will try to find the \code{chrom_list} object used to create the
provided \code{peak_table}.}

\item{what}{`peak_table` or list of chromatograms (`chrom_list`).}

\item{by}{Whether to normalize by a column in sample metadata (\code{meta}) or
by a column in the peak table itself (\code{peak}).}
}
\value{
A \code{peak_table} object where the peaks are normalized by the mass
of each sample.
}
\description{
Normalizes peak table or list of chromatograms by specified column in sample
metadata. Metadata must first be attached to \code{peak_table} using
\code{\link{attach_metadata}}.
}
\examples{
data(pk_tab)
path <- system.file("extdata", "Sa_metadata.csv", package = "chromatographR")
meta <- read.csv(path)
pk_tab <- attach_metadata(peak_table = pk_tab, metadata = meta, column="vial")
norm <- normalize_data(pk_tab, "mass", what = "peak_table")
}
\seealso{
\code{\link{get_peaktable}} \code{\link{attach_metadata}}
}
\author{
Ethan Bass
}
