% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_table-class.R
\name{peak_table-class}
\alias{peak_table-class}
\alias{peak_table}
\title{Peak Table Object}
\description{
S3 objects of class \code{peak_table} store chromatographic peak data along with
associated metadata and reference spectra.
}
\section{Components}{

A \code{peak_table} object contains the following elements:
\itemize{
\item \code{tab}: The peak table itself -- a \code{data.frame} of intensities in a
sample x peak configuration.
\item \code{pk_meta}: A \code{data.frame} containing peak meta-data (e.g., the spectral component,
peak number, and average retention time).
\item \code{sample_meta}: A \code{data.frame} of sample meta-data. Must be added using
\code{\link{attach_metadata}}.
\item \code{ref_spectra}: A \code{data.frame} of reference spectra (in a wavelength x peak
configuration). Must be added using \code{\link{attach_ref_spectra}}.
\item \code{args}: A vector of arguments given to \code{\link{get_peaktable}} to generate
the peak table.
}
}

\section{Methods}{

The following methods are available:
\itemize{
\item \code{\link[=print.peak_table]{print.peak_table()}}
\item \code{\link[=dim.peak_table]{dim.peak_table()}}
\item \code{\link[=head.peak_table]{head.peak_table()}}
\item \code{\link[=tail.peak_table]{tail.peak_table()}}
\item \code{\link[=row.names.peak_table]{row.names.peak_table()}}
\item \code{\link[=subset.peak_table]{subset.peak_table()}}
\item \code{\link[=summary.peak_table]{summary.peak_table()}}
}
}

\examples{
# See get_peaktable() for creation examples
# To access peak_table components:
data(pk_tab)
pk_tab$tab          # peak table
pk_tab$pk_meta      # peak metadata  
pk_tab$sample_meta  # sample metadata

}
\seealso{
\code{\link[=get_peaktable]{get_peaktable()}} for creating peak_table objects
}
