% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatographR-package.R
\docType{data}
\name{pk_tab}
\alias{pk_tab}
\title{Goldenrod peak table}
\format{
A \code{peak_table} object.
}
\description{
Peak table generated from exemplary goldenrod root extracts for use in examples.
}
