% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.peak_list.R
\name{plot.peak_list}
\alias{plot.peak_list}
\title{Plot fitted peak shapes.}
\usage{
\method{plot}{peak_list}(
  x,
  ...,
  chrom_list,
  idx = 1,
  lambda = NULL,
  points = FALSE,
  ticks = FALSE,
  alpha = 0.5,
  color = NULL,
  cex.points = 0.5,
  numbers = FALSE,
  cex.font = 0.5,
  y.offset = 25,
  plot_purity = FALSE,
  res
)
}
\arguments{
\item{x}{A \code{peak_list} object. Output from the \code{get_peaks} function.}

\item{...}{Additional arguments to main plot function.}

\item{chrom_list}{List of chromatograms (retention time x wavelength
matrices)}

\item{idx}{Index or name of chromatogram to be plotted.}

\item{lambda}{Wavelength to use for plotting.}

\item{points}{Logical. If \code{TRUE}, plot peak maxima. Defaults to \code{FALSE}.}

\item{ticks}{Logical. If \code{TRUE}, mark beginning and end of each peak.
Defaults to \code{FALSE}.}

\item{alpha}{Parameter controlling the transparency of fitted shapes.
Defaults to \code{0.5}.}

\item{color}{The color of the fitted shapes.}

\item{cex.points}{Size of points. Defaults to \code{0.5}.}

\item{numbers}{Whether to number peaks. Defaults to \code{FALSE}.}

\item{cex.font}{Font size if peaks are numbered. Defaults to \code{0.5}.}

\item{y.offset}{Y offset for peak numbers. Defaults to \code{25}.}

\item{plot_purity}{Logical. Whether to add visualization of peak purity.
Defaults to \code{FALSE}.}

\item{res}{Time resolution for peak fitting.}
}
\value{
No return value, called for side effects.
}
\description{
Visually assess integration accuracy by plotting fitted peaks over trace.
}
\section{Side effects}{

Plots a chromatographic trace from the specified chromatogram (\code{chr})
at the specified wavelength (\code{lambda}) with fitted peak shapes from the
provided \code{peak_list} drawn underneath the curve.
}

\examples{
data(Sa_warp)
pks <- get_peaks(chrom_list = Sa_warp[1], lambdas = 210)
plot(pks, points = TRUE, ticks = TRUE)
}
\seealso{
\code{\link{get_peaks}}

Other visualization functions:
\code{\link[=boxplot.peak_table]{boxplot.peak_table()}},
\code{\link[=mirror_plot]{mirror_plot()}},
\code{\link[=plot.peak_table]{plot.peak_table()}},
\code{\link[=plot_all_spectra]{plot_all_spectra()}},
\code{\link[=plot_chroms]{plot_chroms()}},
\code{\link[=plot_chroms_heatmap]{plot_chroms_heatmap()}},
\code{\link[=plot_spectrum]{plot_spectrum()}},
\code{\link[=scan_chrom]{scan_chrom()}}
}
\author{
Ethan Bass
}
\concept{visualization functions}
