% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.peak_list.R
\name{plot.peak_list}
\alias{plot.peak_list}
\title{Plot fitted peak shapes.}
\usage{
\method{plot}{peak_list}(
  x,
  ...,
  chrom_list,
  idx = 1,
  lambda = NULL,
  points = FALSE,
  ticks = FALSE,
  a = 0.5,
  color = NULL,
  cex.points = 0.5,
  numbers = FALSE,
  cex.font = 0.5,
  y.offset = 25,
  plot_purity = FALSE,
  res,
  index = NULL
)
}
\arguments{
\item{x}{A \code{peak_list} object. Output from the \code{get_peaks} function.}

\item{...}{Additional arguments to main plot function.}

\item{chrom_list}{List of chromatograms (retention time x wavelength
matrices)}

\item{idx}{Index or name of chromatogram to be plotted.}

\item{lambda}{Wavelength for plotting.}

\item{points}{Logical. If TRUE, plot peak maxima. Defaults to FALSE.}

\item{ticks}{Logical. If TRUE, mark beginning and end of each peak. Defaults
to FALSE.}

\item{a}{Alpha parameter controlling the transparency of fitted shapes.}

\item{color}{The color of the fitted shapes.}

\item{cex.points}{Size of points. Defaults to 0.5}

\item{numbers}{Whether to number peaks. Defaults to FALSE.}

\item{cex.font}{Font size if peaks are numbered. Defaults to 0.5.}

\item{y.offset}{Y offset for peak numbers. Defaults to 25.}

\item{plot_purity}{Whether to add visualization of peak purity.}

\item{res}{time resolution for peak fitting}

\item{index}{This argument is deprecated. Please use \code{idx} instead.}
}
\value{
No return value, called for side effects.
}
\description{
Visually assess integration accuracy by plotting fitted peaks over trace.
}
\section{Side effects}{

Plots a chromatographic trace from the specified chromatogram (\code{chr})
at the specified wavelength (\code{lambda}) with fitted peak shapes from the
provided \code{peak_list} drawn underneath the curve.
}

\seealso{
\code{\link{get_peaks}}

Other visualization functions: 
\code{\link{boxplot.peak_table}()},
\code{\link{mirror_plot}()},
\code{\link{plot.peak_table}()},
\code{\link{plot_all_spectra}()},
\code{\link{plot_chroms}()},
\code{\link{plot_spectrum}()},
\code{\link{scan_chrom}()}
}
\author{
Ethan Bass
}
\concept{visualization functions}
