% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.peak_table.R
\name{plot.peak_table}
\alias{plot.peak_table}
\title{Plot spectrum from peak table}
\usage{
\method{plot}{peak_table}(
  x,
  loc,
  chrom_list,
  what = "peak",
  idx = "max",
  lambda = "max",
  plot_spectrum = TRUE,
  plot_trace = TRUE,
  box_plot = FALSE,
  vars = NULL,
  spectrum_labels = TRUE,
  scale_spectrum = FALSE,
  export_spectrum = FALSE,
  verbose = TRUE,
  engine = c("base", "plotly", "ggplot"),
  chr = NULL,
  ...
)
}
\arguments{
\item{x}{The peak table (output from \code{\link{get_peaktable}}
function).}

\item{loc}{A vector specifying the peak(s) or retention time(s) that you wish to plot.}

\item{chrom_list}{A list of chromatograms in matrix format (timepoints x
wavelengths). If no argument is provided here, the function will try to find the
\code{chrom_list} object used to create the \code{peak_table}.}

\item{what}{What to look for. Either \code{peak} to extract spectral information
for a certain peak, \code{rt} to scan by retention time, or \code{click} to manually
select retention time by clicking on the chromatogram. Defaults to \code{peak}.}

\item{idx}{Numerical index of chromatogram you wish to plot; "max" to
plot the chromatogram with the largest signal; or "all" to plot spectra
for all chromatograms.}

\item{lambda}{The wavelength you wish to plot the trace at (if
\code{plot_chrom} is TRUE and/or the wavelength to be used for the determination
of signal abundance.}

\item{plot_spectrum}{Logical. If TRUE, plots the spectrum of the chosen
peak. Defaults to TRUE.}

\item{plot_trace}{Logical. If TRUE, plots the trace of the chosen peak at
lambda. Defaults to TRUE.}

\item{box_plot}{Logical. If TRUE, plots box plot using categories
defined by \code{vars}.}

\item{vars}{Independent variables for boxplot. Righthand side of formula.}

\item{spectrum_labels}{Logical. If TRUE, plots labels on maxima in spectral
plot. Defaults to TRUE.}

\item{scale_spectrum}{Logical. If TRUE, scales spectrum to unit height.
Defaults to FALSE.}

\item{export_spectrum}{Logical. If TRUE, exports spectrum to console.
Defaults to FALSE.}

\item{verbose}{Logical. If TRUE, prints verbose output to console. Defaults
to TRUE.}

\item{engine}{Which plotting engine to use: either \code{base} or \code{plotly}.}

\item{chr}{Deprecated. Please use \code{idx} instead.}

\item{...}{Additional arguments to \code{\link[graphics]{boxplot}}.}
}
\value{
If \code{export_spectrum} is TRUE, returns the spectrum as a \code{
data.frame} with wavelengths as rows and columns encoding the
absorbance (or normalized absorbance, if \code{scale_spectrum} is TRUE) for 
the specified sample(s). Otherwise, there is no return value.
}
\description{
Plots the trace and/or spectrum for a given peak in peak table.
}
\details{
Can be used to confirm the identity of a peak or check that a particular
column in the peak table represents a single compound. Can also be used
to create simple box-plots to examine the distribution of a peak with respect
to variables defined in sample metadata.
}
\section{Side effects}{

If \code{plot_trace} is \code{TRUE}, plots the chromatographic trace of the 
specified chromatogram (\code{idx}), at the specified wavelength 
(\code{lambda}) with a dotted red line to indicate the retention time given 
by \code{loc}. The trace is a single column from the chromatographic matrix.

If \code{plot_spectrum} is TRUE, plots the spectrum for the specified chromatogram
at the specified retention time. The spectrum is a single row from the chromatographic
matrix.

If \code{box_plot} is TRUE, produces a \code{\link[graphics]{boxplot}} from the
specified peak with groups provided by \code{vars}.
}

\seealso{
Other visualization functions: 
\code{\link{plot.peak_list}()},
\code{\link{plot_all_spectra}()},
\code{\link{plot_chroms}()},
\code{\link{plot_spectrum}()},
\code{\link{scan_chrom}()}
}
\author{
Ethan Bass
}
\concept{visualization functions}
