% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptw_list-methods.R
\name{plot.ptw_list}
\alias{plot.ptw_list}
\title{Plot PTW alignments}
\usage{
\method{plot}{ptw_list}(
  x,
  what = c("traces", "heatmap"),
  engine = c("base", "ggplot", "plotly"),
  lambdas,
  show_legend = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{ptw_list} object created by \code{\link{correct_rt}}.}

\item{what}{What type of plot to return. Either \code{traces} or \code{heatmap}.}

\item{engine}{What plotting engine to use. One of \code{base}, \code{ggplot}, or
\code{plotly}.}

\item{lambdas}{Which lambdas to plot.}

\item{show_legend}{Logical. Whether to include sample legend.}

\item{...}{Additional arguments (placeholder).}
}
\value{
A \code{plotly} or \code{ggplot} object when \code{engine = "plotly"} or
\code{engine = "ggplot"}, respectively. No value is returned when
\code{engine = "base"}.
}
\description{
Plots \code{\link[ptw:ptw]{ptw}} alignments.
}
\details{
Plots PTW alignments at the specified wavelength (\code{lambda}) either as
individual traces or as a heatmap, according to the value of \code{what}. The
plot can be produced using either base R graphics, \code{ggplot2}, or \code{plotly},
according to the value of the \code{engine} argument.
}
\section{Side effects}{

If \code{engine == "base"}, plots are rendered to the active graphics device.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
data(Sa_pr)
warp <- correct_rt(chrom_list = Sa_pr, what = "models", lambdas = 210)
plot(warp)
\dontshow{\}) # examplesIf}
}
\author{
Ethan Bass
}
