% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_peaks.R
\name{plot.ptw_list}
\alias{plot.ptw_list}
\title{Plot PTW alignments}
\usage{
\method{plot}{ptw_list}(
  x,
  what = c("traces", "heatmap"),
  engine = c("base", "ggplot", "plotly"),
  lambdas,
  show_legend = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{ptw_list} object created by \code{\link{correct_rt}}.}

\item{what}{What type of plot to return. Either \code{traces} or 
\code{heatmap}.}

\item{engine}{What plotting engine to use. Either \code{base}, \code{ggplot}
or \code{plotly}}

\item{lambdas}{Which lambdas to plot.}

\item{show_legend}{Logical. Whether to include sample legend.}

\item{...}{Additional arguments (placeholder).}
}
\description{
Plots \code{\link[ptw]{ptw}} alignments.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
data(Sa_pr)
warp <- correct_rt(chrom_list = Sa_pr, what = "models", lambdas = 210)
plot(warp)
\dontshow{\}) # examplesIf}
}
\author{
Ethan Bass
}
