% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_rt.R
\name{plot.ptw_list}
\alias{plot.ptw_list}
\title{Plot PTW alignments}
\usage{
\method{plot}{ptw_list}(x, lambdas, legend = TRUE, ...)
}
\arguments{
\item{x}{A \code{ptw_list} object created by \code{\link{correct_rt}}.}

\item{lambdas}{Which lambdas to plot.}

\item{legend}{Logical. Whether to label the plots.}

\item{...}{Additional arguments to \code{\link[graphics]{matplot}}.}
}
\description{
Plot PTW alignments
}
\author{
Ethan Bass
}
