% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{plot_all_spectra}
\alias{plot_all_spectra}
\title{Plot all spectra for chosen peak}
\usage{
plot_all_spectra(
  loc,
  peak_table,
  chrom_list,
  idx = "all",
  engine = c("base", "ggplot2", "plotly"),
  plot_spectrum = TRUE,
  export_spectrum = TRUE,
  scale_spectrum = TRUE,
  overlapping = TRUE,
  verbose = FALSE,
  what = c("peak", "rt", "idx"),
  peak = NULL,
  ...
)
}
\arguments{
\item{loc}{The peak, retention time, or scan index from which to extract
spectral data.}

\item{peak_table}{A \code{peak_table} object created by \code{\link{get_peaktable}}.}

\item{chrom_list}{A list of chromatograms in matrix format (timepoints x
wavelengths). If no argument is provided here, the function will try to find
the \code{chrom_list} object using the pointer in the \code{peak_table}.}

\item{idx}{Vector of chromatograms to plot.}

\item{engine}{Which plotting engine to use: \code{base}, \code{ggplot2}, or
\code{plotly}.}

\item{plot_spectrum}{Logical. If \code{TRUE}, plots the spectrum of the chosen
peak. Defaults to \code{TRUE}.}

\item{export_spectrum}{Logical. If \code{TRUE}, invisibly returns the spectrum as a \code{data.frame}.
Defaults to \code{FALSE}.}

\item{scale_spectrum}{Logical. If \code{TRUE}, scales spectrum to unit height.
Defaults to \code{FALSE}.}

\item{overlapping}{Logical. If \code{TRUE}, plot spectra in single plot.}

\item{verbose}{Logical. If \code{TRUE}, prints verbose output to console. Defaults
to \code{FALSE}.}

\item{what}{What to look for. Either \code{"peak"} to extract spectral
information for a certain peak, \code{"rt"} to scan by retention time, or
\code{"idx"} to scan by numeric index. Defaults to "peak" mode.}

\item{peak}{The name of a peak to plot (in character format).}

\item{...}{Additional arguments to \code{plot_spectrum.}}
}
\value{
\itemize{
\item If \code{export_spectrum = FALSE} (default), a \code{plotly} or \code{ggplot} object, or
nothing if \code{engine == "base"}.
\item If \code{export_spectrum = TRUE}, invisibly returns a \code{data.frame} with
wavelengths as rows and one column per sample encoding absorbance at each
wavelength (normalized if \code{scale_spectrum = TRUE}).
}
}
\description{
Plot multiple for a given peak in peak table. Wrapper for \code{\link{plot_spectrum}}.
}
\section{Side effects}{

If \code{engine == "base"}, the spectra are rendered to the active graphics
device.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
data(Sa_warp)
pks <- get_peaks(Sa_warp, lambda = "220")
pk_tab <- get_peaktable(pks)
plot_all_spectra(peak = "V13", peak_table = pk_tab, overlapping = TRUE)
\dontshow{\}) # examplesIf}
}
\seealso{
Other visualization functions:
\code{\link[=boxplot.peak_table]{boxplot.peak_table()}},
\code{\link[=mirror_plot]{mirror_plot()}},
\code{\link[=plot.peak_list]{plot.peak_list()}},
\code{\link[=plot.peak_table]{plot.peak_table()}},
\code{\link[=plot_chroms]{plot_chroms()}},
\code{\link[=plot_chroms_heatmap]{plot_chroms_heatmap()}},
\code{\link[=plot_spectrum]{plot_spectrum()}},
\code{\link[=scan_chrom]{scan_chrom()}}
}
\author{
Ethan Bass
}
\concept{visualization functions}
