% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{plot_all_spectra}
\alias{plot_all_spectra}
\title{Plot all spectra for chosen peak.}
\usage{
plot_all_spectra(
  peak,
  peak_table,
  chrom_list,
  idx = "all",
  chrs = NULL,
  engine = c("base", "ggplot2", "plotly"),
  plot_spectrum = TRUE,
  export_spectrum = TRUE,
  scale_spectrum = TRUE,
  overlapping = TRUE,
  verbose = FALSE,
  what = c("peak", "rt", "idx"),
  ...
)
}
\arguments{
\item{peak}{The name of a peak to plot (in character
format).}

\item{peak_table}{The peak table (output from \code{\link{get_peaktable}}
function).}

\item{chrom_list}{A list of chromatograms in matrix format (timepoints x 
components). If no argument is provided here, the function will
try to find the \code{chrom_list} object used to create the provided
\code{peak_table}.}

\item{idx}{Vector of chromatograms to plot.}

\item{chrs}{Deprecated. Please use \code{idx} instead.}

\item{engine}{Which plotting engine to use: \code{base}, \code{ggplot2},
or \code{plotly}.}

\item{plot_spectrum}{Logical. If TRUE, plots the spectrum of the chosen
peak.}

\item{export_spectrum}{Logical. If TRUE, exports spectrum to console.
Defaults to FALSE.}

\item{scale_spectrum}{Logical. If TRUE, scales spectrum to unit height.}

\item{overlapping}{Logical. If TRUE, plot spectra in single plot.}

\item{verbose}{Logical. If TRUE, prints verbose output to console.}

\item{what}{What to look for. Either \code{peak} to extract spectral 
information for a certain peak, \code{rt} to scan by retention time, or 
\code{idx} to scan by numeric index. Defaults to "peak" mode.}

\item{\dots}{Additional arguments to plot_spectrum.}
}
\value{
If \code{export_spectrum} is \code{TRUE}, invisibly returns the 
spectra as a \code{data.frame} with wavelengths as rows and one column for 
each sample in the \code{chrom_list} encoding the absorbance (or normalized 
absorbance, if \code{scale_spectrum} is \code{TRUE}) at each wavelength. 
Otherwise, there is no return value.
}
\description{
Plot multiple for a given peak in peak table. Wrapper for
\code{\link{plot_spectrum}}.
}
\section{Side effects}{

If \code{plot_spectrum} is \code{TRUE}, plots the spectra for the specified 
chromatogram (\code{idx}) of the given \code{peak}. Each spectrum is a single 
row from the chromatographic matrix.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
data(Sa_warp)
pks <- get_peaks(Sa_warp, lambda = "220")
pk_tab <- get_peaktable(pks)
plot_all_spectra(peak = "V13", peak_table = pk_tab, overlapping = TRUE)
\dontshow{\}) # examplesIf}
}
\seealso{
Other visualization functions: 
\code{\link{boxplot.peak_table}()},
\code{\link{mirror_plot}()},
\code{\link{plot.peak_list}()},
\code{\link{plot.peak_table}()},
\code{\link{plot_chroms}()},
\code{\link{plot_chroms_heatmap}()},
\code{\link{plot_spectrum}()},
\code{\link{scan_chrom}()}
}
\author{
Ethan Bass
}
\concept{visualization functions}
