% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chroms.R
\name{plot_chroms}
\alias{plot_chroms}
\title{Plot traces from list of chromatograms.}
\usage{
plot_chroms(
  x,
  lambdas,
  idx,
  time_resolution = 0.01,
  time_unit = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = "",
  ylab = "Absorbance",
  engine = c("base", "ggplot", "plotly"),
  linewidth = 1,
  show_legend = FALSE,
  legend_position = "topright",
  title = "",
  ...
)
}
\arguments{
\item{x}{A list of chromatograms in matrix format (timepoints x wavelengths)
or a \code{peak_table} containing a pointer toward the chromatograms, if present
in the environment.}

\item{lambdas}{A character or numeric vector specifying the wavelengths to
plot. For one-dimensional chromatograms, this argument can be ignored.}

\item{idx}{A vector representing the names or numerical indices of the
chromatograms to plot.}

\item{time_resolution}{Time resolution for plot (in minutes). Defaults to
\code{0.01}. Thinning the time axis dramatically improved speed when plotting
many chromatograms.}

\item{time_unit}{Time units of the provided chromatograms: either \code{min}, \code{s},
or \code{ms}. If possible, units will be detected automatically if possible from
chromatogram metadata. If \code{time_unit} attribute is not present and no
argument is provided, the time units will default to minutes.}

\item{xlim}{Range of x axis (numeric vector).}

\item{ylim}{Range of y axis (numeric vector).}

\item{xlab}{X label.}

\item{ylab}{Y label. Defaults to \code{"Absorbance"}.}

\item{engine}{Plotting engine. Either \code{\link[graphics:matplot]{"base"}},
\code{\link[plotly:plotly]{"plotly"}}, or \code{\link[ggplot2:ggplot2-package]{"ggplot"}}.}

\item{linewidth}{Line width. Defaults to \code{1}.}

\item{show_legend}{Logical. Whether to display legend or not. Defaults to
\code{FALSE}.}

\item{legend_position}{Position of legend. Defaults to \code{"topright"}}

\item{title}{Title for plot.}

\item{...}{Additional arguments to plotting function specified by \code{engine}.}
}
\value{
No return value, called for side effects.
}
\description{
Plots the specified traces from a list of chromatograms.
}
\section{Side effects}{

Plots the traces of the specified chromatograms \code{idx} at the specified
wavelengths (\code{lambdas}). Plots can be produced using base graphics, ggplot2,
or plotly, according to the value of the \code{engine} argument.
}

\examples{
data(Sa_warp)
plot_chroms(Sa_warp, lambdas = 210)
}
\seealso{
Other visualization functions:
\code{\link[=boxplot.peak_table]{boxplot.peak_table()}},
\code{\link[=mirror_plot]{mirror_plot()}},
\code{\link[=plot.peak_list]{plot.peak_list()}},
\code{\link[=plot.peak_table]{plot.peak_table()}},
\code{\link[=plot_all_spectra]{plot_all_spectra()}},
\code{\link[=plot_chroms_heatmap]{plot_chroms_heatmap()}},
\code{\link[=plot_spectrum]{plot_spectrum()}},
\code{\link[=scan_chrom]{scan_chrom()}}
}
\author{
Ethan Bass
}
\concept{visualization functions}
