% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chroms.R
\name{plot_chroms}
\alias{plot_chroms}
\title{Plot traces from list of chromatograms.}
\usage{
plot_chroms(
  x,
  lambdas,
  idx,
  xlim,
  ylim,
  xlab = "",
  ylab = "Absorbance",
  engine = c("base", "ggplot", "plotly"),
  linewidth = 1,
  show_legend = TRUE,
  legend_position = "topright",
  ...
)
}
\arguments{
\item{x}{A list of chromatograms in matrix format (timepoints x wavelengths).}

\item{lambdas}{The wavelength(s) you wish to plot the trace at.}

\item{idx}{A vector representing the names or numerical indices of the 
chromatograms to plot.}

\item{xlim}{Range of x axis.}

\item{ylim}{Range of y axis.}

\item{xlab}{X label.}

\item{ylab}{Y label. Defaults to "Absorbance".}

\item{engine}{Plotting engine. Either \code{\link[graphics]{matplot}}, 
\code{\link[plotly]{plotly}}, or \code{\link[ggplot2]{ggplot2-package}}.}

\item{linewidth}{Line width.}

\item{show_legend}{Logical. Whether to display legend or not. Defaults to TRUE.}

\item{legend_position}{Position of legend.}

\item{...}{Additional arguments to plotting function specified by \code{engine}.}
}
\description{
Plots the specified traces from a list of chromatograms.
}
\examples{
data(Sa_pr)
plot_chroms(Sa_pr, idx=1:2, lambdas=c("210"))
}
