% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chroms.R
\name{plot_chroms_heatmap}
\alias{plot_chroms_heatmap}
\title{Plot chromatograms as heatmap}
\usage{
plot_chroms_heatmap(
  x,
  idx = NULL,
  lambdas,
  engine = c("base", "ggplot", "plotly"),
  show_legend = TRUE,
  xlim = NULL,
  legend_position = "topright",
  title = "",
  show_ylabs = FALSE
)
}
\arguments{
\item{x}{A list of chromatograms in matrix format (timepoints × wavelengths)
or a \code{peak_table} containing a pointer to a list of chromatograms accessible
in the current environment.}

\item{idx}{A vector representing the names or numerical indices of the
chromatograms to plot.}

\item{lambdas}{A character or numeric vector specifying the wavelengths to
plot. Ignored for one-dimensional chromatograms.}

\item{engine}{Plotting backend to use. One of \code{\link[graphics:matplot]{"base"}},
\code{\link[ggplot2:ggplot2-package]{"ggplot"}}, or \code{\link[plotly:plotly]{"plotly"}}.}

\item{show_legend}{Logical. Whether to display legend or not. Defaults to
\code{TRUE}.}

\item{xlim}{Range of x axis values.}

\item{legend_position}{Position of legend. Defaults to \code{"topright"}}

\item{title}{Title for plot.}

\item{show_ylabs}{Logical. Whether to show y labels. Defaults to \code{FALSE}.}
}
\value{
A \code{plotly} or \code{ggplot} object when \code{engine = "plotly"} or
\code{engine = "ggplot"}, respectively. No value is returned when
\code{engine = "base"}.
}
\description{
Plots the specified traces from a list of chromatograms as a heatmap.
}
\details{
Plots the traces of the chromatograms specified by \code{idx} at the specified
wavelengths (\code{lambdas}) as a heatmap. Plots can be produced using base
graphics engine, \code{ggplot2}, or \code{plotly}, according to the value of the
\code{engine} argument. Adapted from \code{\link[VPdtw:plot.VPdtw]{VPdtw::plot.VPdtw}}.
}
\section{Side effects}{

If \code{engine == "base"}, the plot is rendered to the active graphics device.
}

\examples{
data(Sa_warp)
plot_chroms_heatmap(Sa_warp, lambdas = 210)
}
\seealso{
Other visualization functions:
\code{\link[=boxplot.peak_table]{boxplot.peak_table()}},
\code{\link[=mirror_plot]{mirror_plot()}},
\code{\link[=plot.peak_list]{plot.peak_list()}},
\code{\link[=plot.peak_table]{plot.peak_table()}},
\code{\link[=plot_all_spectra]{plot_all_spectra()}},
\code{\link[=plot_chroms]{plot_chroms()}},
\code{\link[=plot_spectrum]{plot_spectrum()}},
\code{\link[=scan_chrom]{scan_chrom()}}
}
\author{
Ethan Bass
}
\concept{visualization functions}
