% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Plot spectrum from peak table}
\usage{
plot_spectrum(
  loc = NULL,
  peak_table,
  chrom_list = NULL,
  idx = "max",
  lambda = "max",
  plot_spectrum = TRUE,
  plot_trace = TRUE,
  spectrum_labels = TRUE,
  scale_spectrum = FALSE,
  export_spectrum = FALSE,
  verbose = TRUE,
  what = c("peak", "rt", "idx", "click"),
  engine = c("base", "plotly", "ggplot2"),
  ...
)
}
\arguments{
\item{loc}{The peak, retention time, or scan index from which to extract
spectral data.}

\item{peak_table}{A \code{peak_table} object created by \code{\link{get_peaktable}}.}

\item{chrom_list}{A list of chromatograms in matrix format (timepoints x
wavelengths). If no argument is provided here, the function will try to find
the \code{chrom_list} object using the pointer in the \code{peak_table}.}

\item{idx}{Numerical index of chromatogram you wish to plot, or "max" to
automatically select the chromatogram with the highest signal intensity at
the specified peak or retention time.}

\item{lambda}{The wavelength used for plotting chromatographic traces and
determining signal intensity.}

\item{plot_spectrum}{Logical. If \code{TRUE}, plots the spectrum of the chosen
peak. Defaults to \code{TRUE}.}

\item{plot_trace}{Logical. If \code{TRUE}, plots the trace of the chosen peak at
lambda. Defaults to \code{TRUE}.}

\item{spectrum_labels}{Logical. If \code{TRUE}, plots labels on maxima in spectral
plot. Defaults to \code{TRUE}.}

\item{scale_spectrum}{Logical. If \code{TRUE}, scales spectrum to unit height.
Defaults to \code{FALSE}.}

\item{export_spectrum}{Logical. If \code{TRUE}, invisibly returns the spectrum as a \code{data.frame}.
Defaults to \code{FALSE}.}

\item{verbose}{Logical. If \code{TRUE} (default), prints verbose output to console.}

\item{what}{What to look for. Either \code{peak} to extract spectral
information for a certain peak, \code{rt} to scan by retention time,
\code{idx} to scan by numeric index, or \code{click} to manually select
retention time by clicking on the chromatogram. Defaults to \code{"peak"} mode.}

\item{engine}{Which plotting engine to use: \code{base}, \code{ggplot2}, or
\code{plotly}.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item If \code{export_spectrum = FALSE} (default), returns a \code{plotly} or \code{ggplot}
object according to the specified \code{engine}. No value is returned when
\code{engine = "base"}.
\item If \code{export_spectrum} is \code{TRUE}, invisibly returns the spectrum as a
\code{data.frame} with wavelengths as rows and a single column encoding the
absorbance (normalized, if \code{scale_spectrum = TRUE}).
}
}
\description{
Visualizes chromatographic traces and/or UV spectra for a selected peak,
retention time, or scan index from a \code{peak_table} or list of chromatograms.
Spectra can also be exported as a \code{data.frame} when \code{export_spectrum = TRUE}.
}
\details{
This function can be used to confirm the identity of a peak or assess whether
a peak table column likely represents a single compound. Retention times may
also be selected interactively by clicking on a chromatographic trace when
\code{what = "click"}.

When \code{plot_trace} is \code{TRUE}, the chromatographic trace for the
specified chromatogram (\code{idx}) is plotted at wavelength \code{lambda}, with a
dotted red line indicating the selected retention time (\code{loc}). The trace
corresponds to a single column of the chromatographic matrix.

When \code{plot_spectrum} is \code{TRUE}, the UV spectrum at the specified retention
time is plotted. The spectrum corresponds to a single row of the
chromatographic matrix.
}
\section{Side effects}{

If \code{engine == "base"}, plots are rendered to the active graphics device.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
data(Sa)
pks <- get_peaks(Sa, lambda = "220.00000")
pk_tab <- get_peaktable(pks)
oldpar <- par(no.readonly = TRUE)
par(mfrow = c(2, 1))
plot_spectrum(loc = "V10", peak_table = pk_tab, what = "peak")
par(oldpar)
\dontshow{\}) # examplesIf}
}
\seealso{
Other visualization functions:
\code{\link[=boxplot.peak_table]{boxplot.peak_table()}},
\code{\link[=mirror_plot]{mirror_plot()}},
\code{\link[=plot.peak_list]{plot.peak_list()}},
\code{\link[=plot.peak_table]{plot.peak_table()}},
\code{\link[=plot_all_spectra]{plot_all_spectra()}},
\code{\link[=plot_chroms]{plot_chroms()}},
\code{\link[=plot_chroms_heatmap]{plot_chroms_heatmap()}},
\code{\link[=scan_chrom]{scan_chrom()}}
}
\author{
Ethan Bass
}
\concept{visualization functions}
