% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess time/wavelength data}
\usage{
preprocess(
  X,
  dim1,
  dim2,
  remove.time.baseline = TRUE,
  spec.smooth = TRUE,
  maxI = NULL,
  interpolate_rows = TRUE,
  interpolate_cols = TRUE,
  cl = 2,
  show_progress = NULL,
  outlier_cutoff = 5/60,
  ...
)
}
\arguments{
\item{X}{A numerical data matrix, or list of data matrices. Missing values
are not allowed. If rownames or colnames attributes are used, they should be
numerical and signify time points and wavelengths, respectively.}

\item{dim1}{A new, usually shorter, set of time points (numerical). The
range of these should not exceed the range of the original time points.}

\item{dim2}{A new, usually shorter, set of wavelengths (numerical). The
range of these should not exceed the range of the original wavelengths.}

\item{remove.time.baseline}{Logical, indicating whether baseline correction
should be done in the time direction, according to
\code{\link[ptw:baseline.corr]{baseline.corr}}. Default is \code{TRUE}.}

\item{spec.smooth}{Logical, indicating whether smoothing should be done in
the spectral direction, according to
\code{\link[stats:smooth.spline]{smooth.spline}}. Default is \code{TRUE}.}

\item{maxI}{if given, the maximum intensity in the matrix is set to this
value.}

\item{interpolate_rows}{Logical. Whether to interpolate along the time axis
(\code{dim1}). Defaults to \code{TRUE}.}

\item{interpolate_cols}{Logical. Whether to interpolate along the spectral
axis (\code{dim2}). Defaults to \code{TRUE}.}

\item{cl}{Either an integer specifying the number of clusters to use for
parallel processing or a cluster object created by
\code{\link[parallel:makeCluster]{makeCluster}}. Defaults to \code{2}. On Windows integer
values will be ignored.}

\item{show_progress}{Logical. Whether to show progress bar. Defaults to
\code{TRUE} if \code{\link[pbapply:pbapply]{pbapply}} is installed.}

\item{outlier_cutoff}{Threshold for excluding chromatograms that end
prematurely. Chromatograms ending more than this value (in seconds) before
the median end time are excluded. Default is 5 seconds. Only applies when
\code{dim1} is not specified.}

\item{...}{Additional arguments to \code{\link[ptw:baseline.corr]{baseline.corr}}.}
}
\value{
The function returns the preprocessed data matrix (or list of
matrices), with row names and column names indicating the time points and
wavelengths, respectively.
}
\description{
Standard pre-processing of response matrices, consisting of a time axis and
a spectral axis (e.g. HPLC-DAD/UV data). For smooth data, like UV-VIS data,
the size of the matrix can be reduced by interpolation. By default,
the data are baseline-corrected in the time direction
(\code{\link[ptw:baseline.corr]{baseline.corr}}) and smoothed in the spectral dimension
using cubic smoothing splines (\code{\link{smooth.spline}}).
}
\note{
Adapted from the \code{preprocess} function in the alsace package by
Ron Wehrens: \url{https://github.com/rwehrens/alsace/blob/master/R/preprocess.R}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
data(Sa)
new.ts <- seq(10,18.66,by=.01) # choose time-points
new.lambdas <- seq(200, 318, by = 2) # choose wavelengths
Sa_pr <- preprocess(Sa[[1]], dim1 = new.ts, dim2 = new.lambdas)
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item Wehrens, R., Bloemberg, T.G., and Eilers P.H.C. 2015. Fast
parametric time warping of peak lists. \emph{Bioinformatics}
31:3063-3065. \doi{10.1093/bioinformatics/btv299}.
\item Wehrens, R., Carvalho, E., Fraser, P.D. 2015. Metabolite profiling in
LC–DAD using multivariate curve resolution: the alsace package for R.
\emph{Metabolomics} 11:1:143-154. \doi{10.1007/s11306-014-0683-5}.
}
}
\author{
Ethan Bass
}
