% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocessing smooth time-wavelength data}
\usage{
preprocess(
  X,
  dim1 = tpoints,
  dim2 = lambdas,
  remove.time.baseline = TRUE,
  spec.smooth = TRUE,
  maxI,
  parallel = TRUE,
  mc.cores = 4,
  ...
)
}
\arguments{
\item{X}{A numerical data matrix, missing values are not allowed. If
rownames or colnames attributes are used, they should be numerical and
signify time points and wavelengths, respectively.}

\item{dim1}{A new, usually shorter, set of time points (numerical). The
range of these should not be outside the range of the original time points,
otherwise the function stops with an error message.}

\item{dim2}{A new, usually shorter, set of wavelengths (numerical). The
range of these should not be outside the range of the original wavelengths,
otherwise the function stops with an error message.}

\item{remove.time.baseline}{Logical, indicating whether baseline correction
should be done in the time direction, according to
\code{\link[ptw:baseline.corr]{baseline.corr}}. Default is TRUE.}

\item{spec.smooth}{Logical, indicating whether smoothing should be done in
the spectral direction, according to
\code{\link[stats:smooth.spline]{smooth.spline}}. Default is TRUE.}

\item{maxI}{if given, the maximum intensity in the matrix is set to this
value.}

\item{parallel}{Logical, indicating whether to use parallel processing.
Defaults to TRUE.}

\item{mc.cores}{How many cores to use for parallel processing. Defaults to 4.}

\item{\dots}{Further optional arguments to
\code{\link[ptw:baseline.corr]{baseline.corr}}.}
}
\value{
The function returns the preprocessed data matrix, with rownames and
colnames indicating the time points and wavelengths, respectively.
}
\description{
Standard preprocessing of response matrices where the first axis is a time
axis, and the second a spectral axis. An example is HPLC-DAD data. For
smooth data, like UV-VIS data, there is the option to decrease the size of
the matrix by interpolation. By default, the data are baseline-corrected in
the time direction and smoothed in the spectral dimension.
}
\examples{

data(Sa)
tpoints <- as.numeric(rownames(Sa[[1]]))
lambdas <- as.numeric(colnames(Sa[[1]]))

## limit retention time and wavelength ranges, and do smoothing and
## baseline correction
new.ts <- seq(1,38,by=.01) # choose time-points
new.lambdas <- seq(200, 400, by = 2) # choose wavelengths
Sa1.processed <-
  preprocess(Sa[[1]], dim1 = new.ts, dim2 = new.lambdas)

plot(tpoints, Sa[[1]][,lambdas == '210'],
     xlim = range(new.ts), type = "l", col = "gray",
     main = "Chromatogram at 210 nm", xlab = "Time (min.)",
     ylab = "")
lines(new.ts, Sa1.processed[,new.lambdas == 210], col = "blue")
legend("topleft", lty = 1, col = c("gray", "blue"), bty = "n",
       legend = c("Original data", "Preprocessed data"))

}
\author{
Ron Wehrens, Ethan Bass
}
