% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_table-methods.R
\name{print.peak_table}
\alias{print.peak_table}
\title{Print peak table}
\usage{
\method{print}{peak_table}(x, ...)
}
\arguments{
\item{x}{A \code{\link{peak_table}} object.}

\item{...}{Additional arguments to \code{\link{print}}.}
}
\value{
Prints and invisible returns the \code{tab} element from the supplied
\code{peak_table}.
}
\description{
Print peak table
}
\keyword{internal}
