% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_chroms.R
\docType{import}
\name{read_chroms}
\alias{read_chroms}
\title{Read chromatograms.}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{chromConverter}{\code{\link[chromConverter]{read_chroms}}}
}}

