% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_chroms.R
\name{reshape_chroms}
\alias{reshape_chroms}
\title{Reshape chromatograms}
\usage{
reshape_chroms(x, idx, sample_var = "sample", lambdas = NULL, rts = NULL)
}
\arguments{
\item{x}{A list of chromatographic matrices in wide format.}

\item{idx}{Indices of chromatograms to convert.}

\item{sample_var}{String with name of new column containing sample IDs.}

\item{lambdas}{Vector specifying wavelength(s) to include.}

\item{rts}{Vector specifying retention times to include.}
}
\value{
A list of chromatographic matrices in long format.
}
\description{
Reshapes a list of chromatograms from wide to long format.
}
\seealso{
Other utility functions: 
\code{\link{combine_peaks}()},
\code{\link{get_lambdas}()},
\code{\link{get_times}()},
\code{\link{merge_peaks}()},
\code{\link{reshape_peaktable}()}
}
\author{
Ethan Bass
}
\concept{utility functions}
