% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_chroms.R
\name{reshape_peaktable}
\alias{reshape_peaktable}
\title{Reshapes peak table from wide to long format}
\usage{
reshape_peaktable(x, peaks)
}
\arguments{
\item{x}{A \code{peak_table} object.}

\item{peaks}{A character vector specifying peaks to include.}
}
\value{
A data.frame containing the information for the specified peaks in
long format.
}
\description{
Reshapes peak table from wide to long format
}
\author{
Ethan Bass
}
