\name{Sa}
\alias{Sa}
\docType{data}
\title{HLPC-DAD data of goldenrod root extracts.}
\description{
Five HPLC-DAD data matrices of *Solidago altissima* roots extracted in 90 percent methanol. Since the raw data are given (no smoothing or baseline subtraction has been done, only subsetting of the time and wavelength axes).
}
\usage{data(Sa)}
\format{The UV-Vis data (\code{Sa}) are given as a list of five
  matrices (time x wavelength).
}
\keyword{datasets}
