% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatographR-package.R
\docType{data}
\name{Sa}
\alias{Sa}
\title{Raw goldenrod root chromatograms}
\format{
A list of four matrices (1301 times x 60 wavelengths).
}
\usage{
data(Sa)
}
\description{
A list of four HPLC-DAD data matrices of \emph{Solidago altissima} roots
extracted in 90\% methanol. Retention times are stored in rows and 
wavelengths are stored in columns.
}
\keyword{data}
