% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatographR-package.R
\docType{data}
\name{Sa}
\alias{Sa}
\title{Raw goldenrod root chromatograms}
\format{
A list of four matrices (time x wavelength).
}
\description{
Four HPLC-DAD data matrices of *Solidago altissima* roots extracted in 90%
methanol. The time axis is compressed to save space so the
data are a little choppy.
}
