% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatographR-package.R
\docType{data}
\name{Sa}
\alias{Sa}
\title{HPLC-DAD data of goldenrod root extracts.}
\format{
A list of four matrices (time x wavelength).
}
\description{
Four HPLC-DAD data matrices of *Solidago altissima* roots extracted in 90
percent methanol.
}
