% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatographR-package.R
\docType{data}
\name{Sa_pr}
\alias{Sa_pr}
\title{Preprocessed goldenrod root chromatograms}
\format{
Four pre-processed matrices (time x wavelength) to use in examples.
}
\description{
These are the pre-processed chromatograms derived from the raw data stored in
\code{Sa}.
}
