% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatographR-package.R
\docType{data}
\name{Sa_pr}
\alias{Sa_pr}
\title{HPLC-DAD data of goldenrod root extracts.}
\format{
Four pre-processed matrices (time x wavelength) to use in examples.
}
\description{
Pre-processed chromatograms.
}
