% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatographR-package.R
\docType{data}
\name{Sa_warp}
\alias{Sa_warp}
\title{Warped goldenrod root chromatograms.}
\format{
Four pre-processed and warped matrices (time x wavelength) to use in
examples.
}
\description{
These are pre-processed and warped goldenrod root chromatograms derived from
the raw data stored in \code{Sa}.
}
