% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{scan_chrom}
\alias{scan_chrom}
\title{Scan spectrum}
\usage{
scan_chrom(
  chrom_list,
  lambda,
  chr = NULL,
  peak_table = NULL,
  scale_spectrum = FALSE,
  spectrum_labels = TRUE,
  export_spectrum = FALSE,
  ...
)
}
\arguments{
\item{chrom_list}{A list of chromatograms in matrix form (timepoints x
wavelengths).}

\item{lambda}{The wavelength to plot the trace at.}

\item{chr}{Numerical index of chromatogram you wish to plot.}

\item{peak_table}{The peak table (output from \code{\link{get_peaktable}}
function).}

\item{scale_spectrum}{Logical. If TRUE, scales spectrum to unit height.
Defaults to FALSE.}

\item{spectrum_labels}{Logical. If TRUE, plots labels on maxima in spectral
plot. Defaults to TRUE.}

\item{export_spectrum}{Logical. If TRUE, exports spectrum to console.
Defaults to FALSE.}

\item{...}{Additional arguments.}
}
\description{
Convenience function to call plot_spectrum with what = "click".
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(Sa_short_pr)
scan_chrom(Sa_short_pr, lambda="210", chr=2, export_spectrum=T)
\dontshow{\}) # examplesIf}
}
\author{
Ethan Bass
}
