% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{scan_chrom}
\alias{scan_chrom}
\title{Plot spectra by clicking on the chromatogram}
\usage{
scan_chrom(
  chrom_list,
  idx,
  lambda,
  plot_spectrum = TRUE,
  peak_table = NULL,
  scale_spectrum = FALSE,
  spectrum_labels = TRUE,
  export_spectrum = FALSE,
  chr = NULL,
  ...
)
}
\arguments{
\item{chrom_list}{A list of chromatograms in matrix format (timepoints x
wavelengths). If no argument is provided here, the function will try to find
the \code{chrom_list} object used to create the provided \code{peak_table}.}

\item{idx}{Numerical index of chromatogram you wish to plot.}

\item{lambda}{The wavelength to plot the trace at.}

\item{plot_spectrum}{Logical. Whether to plot the spectrum or not.}

\item{peak_table}{The peak table (output from \code{\link{get_peaktable}}
function).}

\item{scale_spectrum}{Logical. If TRUE, scales spectrum to unit height.
Defaults to FALSE.}

\item{spectrum_labels}{Logical. If TRUE, plots labels on maxima in spectral
plot. Defaults to TRUE.}

\item{export_spectrum}{Logical. If TRUE, exports spectrum to console.
Defaults to FALSE.}

\item{chr}{Deprecated. Please use \code{idx} instead.}

\item{...}{Additional arguments.}
}
\value{
If \code{export_spectrum} is TRUE, returns the spectrum as a \code{
data.frame} with wavelengths as rows and a single column encoding the
absorbance (or normalized absorbance, if \code{scale_spectrum} is TRUE)
at each wavelength. Otherwise, there is no return value.
}
\description{
Plot spectra by clicking on the chromatogram
}
\section{Side effects}{

Plots a chromatographic trace from the specified chromatogram (\code{idx}),
at the specified wavelength (\code{lambda}) with a dotted red line to indicate
the user-selected retention time. The trace is a single column from the
chromatographic matrix.

If \code{plot_spectrum} is TRUE, plots the spectrum for the specified
chromatogram at the user-specified retention time. The spectrum is a single
row from the chromatographic matrix.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
data(Sa_pr)
scan_chrom(Sa_pr, lambda = "210", idx = 2, export_spectrum = TRUE)
\dontshow{\}) # examplesIf}
}
\seealso{
Other visualization functions: 
\code{\link{boxplot.peak_table}()},
\code{\link{mirror_plot}()},
\code{\link{plot.peak_list}()},
\code{\link{plot.peak_table}()},
\code{\link{plot_all_spectra}()},
\code{\link{plot_chroms}()},
\code{\link{plot_spectrum}()}
}
\author{
Ethan Bass
}
\concept{visualization functions}
