% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaktable.R
\name{subset.peak_table}
\alias{subset.peak_table}
\title{Subset peak table}
\usage{
\method{subset}{peak_table}(x, subset, select, drop = FALSE, ...)
}
\arguments{
\item{x}{A \code{peak_table} object.}

\item{subset}{Logical expression indicating rows (samples) to keep from
\code{peak_table}; missing values are taken as false.}

\item{select}{Logical expression indicating columns (peaks) to select from
\code{peak_table}.}

\item{drop}{Logical. Passed to indexing operator.}

\item{...}{Additional arguments (placeholder).}
}
\value{
A \code{peak_table} object with samples specified by \code{subset}
and peaks specified by \code{select}.
}
\description{
Returns subset of \code{peak_table} object.
}
\author{
Ethan Bass
}
