% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_table-methods.R
\name{summary.peak_table}
\alias{summary.peak_table}
\title{Summarize Peak Table}
\usage{
\method{summary}{peak_table}(object, ...)
}
\arguments{
\item{object}{A \code{\link{peak_table}} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A list containing basic information about the supplied 
\code{peak_table} object.
}
\description{
Prints basic statistics about \code{peak_table} to the console (e.g., the 
number of peaks, wavelengths, range of intensities, etc).
}
\author{
Ethan Bass
}
\keyword{internal}
