% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_purity.R
\name{trim_peak}
\alias{trim_peak}
\title{Trim peak}
\usage{
trim_peak(x, pos, cutoff = 0.05)
}
\arguments{
\item{x}{A chromatogram in matrix format.}

\item{pos}{A vector containing peak information.}

\item{cutoff}{Proportion of maximum absorbance to use as cutoff. Defaults to
\code{.05}.}
}
\value{
Returns indices within the peak specified by \code{pos} with a higher
signal intensity than the specified cutoff.
}
\description{
Trim peak
}
\author{
Ethan Bass
}
\keyword{internal}
