% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_purity.R, R/peak_purity.R
\name{trim_peak}
\alias{trim_peak}
\title{Trim peak}
\usage{
trim_peak(x, pos, cutoff = 0.05)

trim_peak(x, pos, cutoff = 0.05)
}
\arguments{
\item{x}{A chromatogram in matrix format}

\item{pos}{A vector containing peak information}

\item{cutoff}{Proportion of maximum absorbance to use as cutoff.}
}
\description{
Trim peak

Trim peak
}
\author{
Ethan Bass
}
