% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_peaktable.R
\name{write_peaktable}
\alias{write_peaktable}
\title{Export peak table}
\usage{
write_peaktable(
  peak_table,
  path,
  filename = "peak_table",
  format = c("csv", "xlsx"),
  what = c("tab", "pk_meta", "sample_meta", "ref_spectra", "args")
)
}
\arguments{
\item{peak_table}{Peak table object from \code{\link{get_peaktable}}.}

\item{path}{Path to write file.}

\item{filename}{File name. Defaults to \code{"peak_table"}.}

\item{format}{File format to export. Either \code{csv} or \code{xlsx}.}

\item{what}{Which elements of the \code{peak_table} to export.}
}
\value{
No return value. The function is called for its side effects.
}
\description{
Exports peak table in \code{csv} or \code{xlsx} format.
}
\section{Side effects}{

Exports peak_table object as \code{.csv} or \code{.xlsx} file according to the value
of the \code{format} argument.
}

\examples{
\donttest{
data(pk_tab)
path_out = tempdir()
write_peaktable(pk_tab, path = path_out, what = c("tab"))
}
}
