% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_peaktable.R
\name{write_peaktable}
\alias{write_peaktable}
\title{Export peak table}
\usage{
write_peaktable(
  peak_table,
  path,
  format = c("csv", "xlsx"),
  what = c("tab", "pk_meta", "sample_meta", "ref_spectra", "args")
)
}
\arguments{
\item{peak_table}{Peak table object from}

\item{path}{path to write file}

\item{format}{format for export. Either \code{csv} or \code{xlsx}.}

\item{what}{which elements of the \code{peak_table} to export.}
}
\description{
Export peak table
}
